Module aws_sdk_ssm::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AddTagsToResource

Fluent builder constructing a request to AddTagsToResource.

AssociateOpsItemRelatedItem

Fluent builder constructing a request to AssociateOpsItemRelatedItem.

CancelCommand

Fluent builder constructing a request to CancelCommand.

CancelMaintenanceWindowExecution

Fluent builder constructing a request to CancelMaintenanceWindowExecution.

CreateActivation

Fluent builder constructing a request to CreateActivation.

CreateAssociation

Fluent builder constructing a request to CreateAssociation.

CreateAssociationBatch

Fluent builder constructing a request to CreateAssociationBatch.

CreateDocument

Fluent builder constructing a request to CreateDocument.

CreateMaintenanceWindow

Fluent builder constructing a request to CreateMaintenanceWindow.

CreateOpsItem

Fluent builder constructing a request to CreateOpsItem.

CreateOpsMetadata

Fluent builder constructing a request to CreateOpsMetadata.

CreatePatchBaseline

Fluent builder constructing a request to CreatePatchBaseline.

CreateResourceDataSync

Fluent builder constructing a request to CreateResourceDataSync.

DeleteActivation

Fluent builder constructing a request to DeleteActivation.

DeleteAssociation

Fluent builder constructing a request to DeleteAssociation.

DeleteDocument

Fluent builder constructing a request to DeleteDocument.

DeleteInventory

Fluent builder constructing a request to DeleteInventory.

DeleteMaintenanceWindow

Fluent builder constructing a request to DeleteMaintenanceWindow.

DeleteOpsMetadata

Fluent builder constructing a request to DeleteOpsMetadata.

DeleteParameter

Fluent builder constructing a request to DeleteParameter.

DeleteParameters

Fluent builder constructing a request to DeleteParameters.

DeletePatchBaseline

Fluent builder constructing a request to DeletePatchBaseline.

DeleteResourceDataSync

Fluent builder constructing a request to DeleteResourceDataSync.

DeregisterManagedInstance

Fluent builder constructing a request to DeregisterManagedInstance.

DeregisterPatchBaselineForPatchGroup

Fluent builder constructing a request to DeregisterPatchBaselineForPatchGroup.

DeregisterTargetFromMaintenanceWindow

Fluent builder constructing a request to DeregisterTargetFromMaintenanceWindow.

DeregisterTaskFromMaintenanceWindow

Fluent builder constructing a request to DeregisterTaskFromMaintenanceWindow.

DescribeActivations

Fluent builder constructing a request to DescribeActivations.

DescribeAssociation

Fluent builder constructing a request to DescribeAssociation.

DescribeAssociationExecutionTargets

Fluent builder constructing a request to DescribeAssociationExecutionTargets.

DescribeAssociationExecutions

Fluent builder constructing a request to DescribeAssociationExecutions.

DescribeAutomationExecutions

Fluent builder constructing a request to DescribeAutomationExecutions.

DescribeAutomationStepExecutions

Fluent builder constructing a request to DescribeAutomationStepExecutions.

DescribeAvailablePatches

Fluent builder constructing a request to DescribeAvailablePatches.

DescribeDocument

Fluent builder constructing a request to DescribeDocument.

DescribeDocumentPermission

Fluent builder constructing a request to DescribeDocumentPermission.

DescribeEffectiveInstanceAssociations

Fluent builder constructing a request to DescribeEffectiveInstanceAssociations.

DescribeEffectivePatchesForPatchBaseline

Fluent builder constructing a request to DescribeEffectivePatchesForPatchBaseline.

DescribeInstanceAssociationsStatus

Fluent builder constructing a request to DescribeInstanceAssociationsStatus.

DescribeInstanceInformation

Fluent builder constructing a request to DescribeInstanceInformation.

DescribeInstancePatchStates

Fluent builder constructing a request to DescribeInstancePatchStates.

DescribeInstancePatchStatesForPatchGroup

Fluent builder constructing a request to DescribeInstancePatchStatesForPatchGroup.

DescribeInstancePatches

Fluent builder constructing a request to DescribeInstancePatches.

DescribeInventoryDeletions

Fluent builder constructing a request to DescribeInventoryDeletions.

DescribeMaintenanceWindowExecutionTaskInvocations

Fluent builder constructing a request to DescribeMaintenanceWindowExecutionTaskInvocations.

DescribeMaintenanceWindowExecutionTasks

Fluent builder constructing a request to DescribeMaintenanceWindowExecutionTasks.

DescribeMaintenanceWindowExecutions

Fluent builder constructing a request to DescribeMaintenanceWindowExecutions.

DescribeMaintenanceWindowSchedule

Fluent builder constructing a request to DescribeMaintenanceWindowSchedule.

DescribeMaintenanceWindowTargets

Fluent builder constructing a request to DescribeMaintenanceWindowTargets.

DescribeMaintenanceWindowTasks

Fluent builder constructing a request to DescribeMaintenanceWindowTasks.

DescribeMaintenanceWindows

Fluent builder constructing a request to DescribeMaintenanceWindows.

DescribeMaintenanceWindowsForTarget

Fluent builder constructing a request to DescribeMaintenanceWindowsForTarget.

DescribeOpsItems

Fluent builder constructing a request to DescribeOpsItems.

DescribeParameters

Fluent builder constructing a request to DescribeParameters.

DescribePatchBaselines

Fluent builder constructing a request to DescribePatchBaselines.

DescribePatchGroupState

Fluent builder constructing a request to DescribePatchGroupState.

DescribePatchGroups

Fluent builder constructing a request to DescribePatchGroups.

DescribePatchProperties

Fluent builder constructing a request to DescribePatchProperties.

DescribeSessions

Fluent builder constructing a request to DescribeSessions.

DisassociateOpsItemRelatedItem

Fluent builder constructing a request to DisassociateOpsItemRelatedItem.

GetAutomationExecution

Fluent builder constructing a request to GetAutomationExecution.

GetCalendarState

Fluent builder constructing a request to GetCalendarState.

GetCommandInvocation

Fluent builder constructing a request to GetCommandInvocation.

GetConnectionStatus

Fluent builder constructing a request to GetConnectionStatus.

GetDefaultPatchBaseline

Fluent builder constructing a request to GetDefaultPatchBaseline.

GetDeployablePatchSnapshotForInstance

Fluent builder constructing a request to GetDeployablePatchSnapshotForInstance.

GetDocument

Fluent builder constructing a request to GetDocument.

GetInventory

Fluent builder constructing a request to GetInventory.

GetInventorySchema

Fluent builder constructing a request to GetInventorySchema.

GetMaintenanceWindow

Fluent builder constructing a request to GetMaintenanceWindow.

GetMaintenanceWindowExecution

Fluent builder constructing a request to GetMaintenanceWindowExecution.

GetMaintenanceWindowExecutionTask

Fluent builder constructing a request to GetMaintenanceWindowExecutionTask.

GetMaintenanceWindowExecutionTaskInvocation

Fluent builder constructing a request to GetMaintenanceWindowExecutionTaskInvocation.

GetMaintenanceWindowTask

Fluent builder constructing a request to GetMaintenanceWindowTask.

GetOpsItem

Fluent builder constructing a request to GetOpsItem.

GetOpsMetadata

Fluent builder constructing a request to GetOpsMetadata.

GetOpsSummary

Fluent builder constructing a request to GetOpsSummary.

GetParameter

Fluent builder constructing a request to GetParameter.

GetParameterHistory

Fluent builder constructing a request to GetParameterHistory.

GetParameters

Fluent builder constructing a request to GetParameters.

GetParametersByPath

Fluent builder constructing a request to GetParametersByPath.

GetPatchBaseline

Fluent builder constructing a request to GetPatchBaseline.

GetPatchBaselineForPatchGroup

Fluent builder constructing a request to GetPatchBaselineForPatchGroup.

GetServiceSetting

Fluent builder constructing a request to GetServiceSetting.

LabelParameterVersion

Fluent builder constructing a request to LabelParameterVersion.

ListAssociationVersions

Fluent builder constructing a request to ListAssociationVersions.

ListAssociations

Fluent builder constructing a request to ListAssociations.

ListCommandInvocations

Fluent builder constructing a request to ListCommandInvocations.

ListCommands

Fluent builder constructing a request to ListCommands.

ListComplianceItems

Fluent builder constructing a request to ListComplianceItems.

ListComplianceSummaries

Fluent builder constructing a request to ListComplianceSummaries.

ListDocumentMetadataHistory

Fluent builder constructing a request to ListDocumentMetadataHistory.

ListDocumentVersions

Fluent builder constructing a request to ListDocumentVersions.

ListDocuments

Fluent builder constructing a request to ListDocuments.

ListInventoryEntries

Fluent builder constructing a request to ListInventoryEntries.

ListOpsItemEvents

Fluent builder constructing a request to ListOpsItemEvents.

ListOpsItemRelatedItems

Fluent builder constructing a request to ListOpsItemRelatedItems.

ListOpsMetadata

Fluent builder constructing a request to ListOpsMetadata.

ListResourceComplianceSummaries

Fluent builder constructing a request to ListResourceComplianceSummaries.

ListResourceDataSync

Fluent builder constructing a request to ListResourceDataSync.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ModifyDocumentPermission

Fluent builder constructing a request to ModifyDocumentPermission.

PutComplianceItems

Fluent builder constructing a request to PutComplianceItems.

PutInventory

Fluent builder constructing a request to PutInventory.

PutParameter

Fluent builder constructing a request to PutParameter.

RegisterDefaultPatchBaseline

Fluent builder constructing a request to RegisterDefaultPatchBaseline.

RegisterPatchBaselineForPatchGroup

Fluent builder constructing a request to RegisterPatchBaselineForPatchGroup.

RegisterTargetWithMaintenanceWindow

Fluent builder constructing a request to RegisterTargetWithMaintenanceWindow.

RegisterTaskWithMaintenanceWindow

Fluent builder constructing a request to RegisterTaskWithMaintenanceWindow.

RemoveTagsFromResource

Fluent builder constructing a request to RemoveTagsFromResource.

ResetServiceSetting

Fluent builder constructing a request to ResetServiceSetting.

ResumeSession

Fluent builder constructing a request to ResumeSession.

SendAutomationSignal

Fluent builder constructing a request to SendAutomationSignal.

SendCommand

Fluent builder constructing a request to SendCommand.

StartAssociationsOnce

Fluent builder constructing a request to StartAssociationsOnce.

StartAutomationExecution

Fluent builder constructing a request to StartAutomationExecution.

StartChangeRequestExecution

Fluent builder constructing a request to StartChangeRequestExecution.

StartSession

Fluent builder constructing a request to StartSession.

StopAutomationExecution

Fluent builder constructing a request to StopAutomationExecution.

TerminateSession

Fluent builder constructing a request to TerminateSession.

UnlabelParameterVersion

Fluent builder constructing a request to UnlabelParameterVersion.

UpdateAssociation

Fluent builder constructing a request to UpdateAssociation.

UpdateAssociationStatus

Fluent builder constructing a request to UpdateAssociationStatus.

UpdateDocument

Fluent builder constructing a request to UpdateDocument.

UpdateDocumentDefaultVersion

Fluent builder constructing a request to UpdateDocumentDefaultVersion.

UpdateDocumentMetadata

Fluent builder constructing a request to UpdateDocumentMetadata.

UpdateMaintenanceWindow

Fluent builder constructing a request to UpdateMaintenanceWindow.

UpdateMaintenanceWindowTarget

Fluent builder constructing a request to UpdateMaintenanceWindowTarget.

UpdateMaintenanceWindowTask

Fluent builder constructing a request to UpdateMaintenanceWindowTask.

UpdateManagedInstanceRole

Fluent builder constructing a request to UpdateManagedInstanceRole.

UpdateOpsItem

Fluent builder constructing a request to UpdateOpsItem.

UpdateOpsMetadata

Fluent builder constructing a request to UpdateOpsMetadata.

UpdatePatchBaseline

Fluent builder constructing a request to UpdatePatchBaseline.

UpdateResourceDataSync

Fluent builder constructing a request to UpdateResourceDataSync.

UpdateServiceSetting

Fluent builder constructing a request to UpdateServiceSetting.