Struct aws_sdk_ssm::client::fluent_builders::CreateMaintenanceWindow

pub struct CreateMaintenanceWindow<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateMaintenanceWindow<C>

pub async fn send(
self
) -> Result<CreateMaintenanceWindowOutput, SdkError<CreateMaintenanceWindowError>> where
C: SmithyConnector,

pub fn name(self, input: impl Into<String>) -> Self

The name of the maintenance window.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

An optional description for the maintenance window. We recommend specifying a description to help you organize your maintenance windows.

pub fn set_description(self, input: Option<String>) -> Self

pub fn start_date(self, input: impl Into<String>) -> Self

The date and time, in ISO-8601 Extended format, for when you want the maintenance window to become active. StartDate allows you to delay activation of the maintenance window until the specified future date.

pub fn set_start_date(self, input: Option<String>) -> Self

pub fn end_date(self, input: impl Into<String>) -> Self

The date and time, in ISO-8601 Extended format, for when you want the maintenance window to become inactive. EndDate allows you to set a date and time in the future when the maintenance window will no longer run.

pub fn set_end_date(self, input: Option<String>) -> Self

pub fn schedule(self, input: impl Into<String>) -> Self

The schedule of the maintenance window in the form of a cron or rate expression.

pub fn set_schedule(self, input: Option<String>) -> Self

pub fn schedule_timezone(self, input: impl Into<String>) -> Self

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

pub fn set_schedule_timezone(self, input: Option<String>) -> Self

pub fn schedule_offset(self, input: i32) -> Self

The number of days to wait after the date and time specified by a CRON expression before running the maintenance window.

For example, the following cron expression schedules a maintenance window to run on the third Tuesday of every month at 11:30 PM.

cron(30 23 ? * TUE#3 *)

If the schedule offset is 2, the maintenance window won't run until two days later.

pub fn set_schedule_offset(self, input: Option<i32>) -> Self

pub fn duration(self, input: i32) -> Self

The duration of the maintenance window in hours.

pub fn set_duration(self, input: Option<i32>) -> Self

pub fn cutoff(self, input: i32) -> Self

The number of hours before the end of the maintenance window that Systems Manager stops scheduling new tasks for execution.

pub fn set_cutoff(self, input: Option<i32>) -> Self

pub fn allow_unassociated_targets(self, input: bool) -> Self

Enables a maintenance window task to run on managed instances, even if you have not registered those instances as targets. If enabled, then you must specify the unregistered instances (by instance ID) when you register a task with the maintenance window.

If you don't enable this option, then you must specify previously-registered targets when you register a task with the maintenance window.

pub fn set_allow_unassociated_targets(self, input: Option<bool>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

User-provided idempotency token.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a maintenance window to identify the type of tasks it will run, the types of targets, and the environment it will run in. In this case, you could specify the following key name/value pairs:

Key=TaskType,Value=AgentUpdate

Key=OS,Value=Windows

Key=Environment,Value=Production

To add tags to an existing maintenance window, use the AddTagsToResource action.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self