Struct aws_sdk_ssm::client::fluent_builders::CreatePatchBaseline

pub struct CreatePatchBaseline<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreatePatchBaseline<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreatePatchBaselineOutput, SdkError<CreatePatchBaselineError>> where
R::Policy: SmithyRetryPolicy<CreatePatchBaselineInputOperationOutputAlias, CreatePatchBaselineOutput, CreatePatchBaselineError, CreatePatchBaselineInputOperationRetryAlias>,

pub fn operating_system(self, inp: OperatingSystem) -> Self

Defines the operating system the patch baseline applies to. The default value is WINDOWS.

pub fn set_operating_system(self, input: Option<OperatingSystem>) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

The name of the patch baseline.

pub fn set_name(self, input: Option<String>) -> Self

pub fn global_filters(self, inp: PatchFilterGroup) -> Self

A set of global filters used to include patches in the baseline.

pub fn set_global_filters(self, input: Option<PatchFilterGroup>) -> Self

pub fn approval_rules(self, inp: PatchRuleGroup) -> Self

A set of rules used to include patches in the baseline.

pub fn set_approval_rules(self, input: Option<PatchRuleGroup>) -> Self

pub fn approved_patches(self, inp: impl Into<String>) -> Self

Appends an item to ApprovedPatches.

To override the contents of this collection use set_approved_patches.

A list of explicitly approved patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

pub fn set_approved_patches(self, input: Option<Vec<String>>) -> Self

pub fn approved_patches_compliance_level(
self,
inp: PatchComplianceLevel
) -> Self

Defines the compliance level for approved patches. When an approved patch is reported as missing, this value describes the severity of the compliance violation. The default value is UNSPECIFIED.

pub fn set_approved_patches_compliance_level(
self,
input: Option<PatchComplianceLevel>
) -> Self

pub fn approved_patches_enable_non_security(self, inp: bool) -> Self

Indicates whether the list of approved patches includes non-security updates that should be applied to the instances. The default value is false. Applies to Linux instances only.

pub fn set_approved_patches_enable_non_security(
self,
input: Option<bool>
) -> Self

pub fn rejected_patches(self, inp: impl Into<String>) -> Self

Appends an item to RejectedPatches.

To override the contents of this collection use set_rejected_patches.

A list of explicitly rejected patches for the baseline.

For information about accepted formats for lists of approved patches and rejected patches, see About package name formats for approved and rejected patch lists in the Amazon Web Services Systems Manager User Guide.

pub fn set_rejected_patches(self, input: Option<Vec<String>>) -> Self

pub fn rejected_patches_action(self, inp: PatchAction) -> Self

The action for Patch Manager to take on patches included in the RejectedPackages list.

ALLOW_AS_DEPENDENCY : A package in the Rejected patches list is installed only if it is a dependency of another package. It is considered compliant with the patch baseline, and its status is reported as InstalledOther. This is the default action if no option is specified.

BLOCK : Packages in the RejectedPatches list, and packages that include them as dependencies, aren't installed under any circumstances. If a package was installed before it was added to the Rejected patches list, it is considered non-compliant with the patch baseline, and its status is reported as InstalledRejected.

pub fn set_rejected_patches_action(self, input: Option<PatchAction>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the patch baseline.

pub fn set_description(self, input: Option<String>) -> Self

pub fn sources(self, inp: impl Into<PatchSource>) -> Self

Appends an item to Sources.

To override the contents of this collection use set_sources.

Information about the patches to use to update the instances, including target operating systems and source repositories. Applies to Linux instances only.

pub fn set_sources(self, input: Option<Vec<PatchSource>>) -> Self

pub fn client_token(self, inp: impl Into<String>) -> Self

User-provided idempotency token.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag a patch baseline to identify the severity level of patches it specifies and the operating system family it applies to. In this case, you could specify the following key-value pairs:

Key=PatchSeverity,Value=Critical

Key=OS,Value=Windows

To add tags to an existing patch baseline, use the AddTagsToResource operation.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self