Struct aws_sdk_ssm::client::fluent_builders::RegisterTaskWithMaintenanceWindow

pub struct RegisterTaskWithMaintenanceWindow<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RegisterTaskWithMaintenanceWindow<C>

pub async fn send(
self
) -> Result<RegisterTaskWithMaintenanceWindowOutput, SdkError<RegisterTaskWithMaintenanceWindowError>> where
C: SmithyConnector,

pub fn window_id(self, input: impl Into<String>) -> Self

The ID of the maintenance window the task should be added to.

pub fn set_window_id(self, input: Option<String>) -> Self

pub fn targets(self, inp: impl Into<Target>) -> Self

The targets (either instances or maintenance window targets).

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, AWS Lambda, and AWS Step Functions). For more information about running tasks that do not specify targets, see Registering maintenance window tasks without targets in the AWS Systems Manager User Guide.

Specify instances using the following format:

Key=InstanceIds,Values=,

Specify maintenance window targets using the following format:

Key=WindowTargetIds,Values=,

pub fn set_targets(self, input: Option<Vec<Target>>) -> Self

pub fn task_arn(self, input: impl Into<String>) -> Self

The ARN of the task to run.

pub fn set_task_arn(self, input: Option<String>) -> Self

pub fn service_role_arn(self, input: impl Into<String>) -> Self

The ARN of the IAM service role for Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the AWS Systems Manager User Guide:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

pub fn set_service_role_arn(self, input: Option<String>) -> Self

pub fn task_type(self, input: MaintenanceWindowTaskType) -> Self

The type of task being registered.

pub fn set_task_type(self, input: Option<MaintenanceWindowTaskType>) -> Self

pub fn task_parameters(
self,
k: impl Into<String>,
v: impl Into<MaintenanceWindowTaskParameterValueExpression>
) -> Self

The parameters that should be passed to the task when it is run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

pub fn set_task_parameters(
self,
input: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>
) -> Self

pub fn task_invocation_parameters(
self,
input: MaintenanceWindowTaskInvocationParameters
) -> Self

The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

pub fn set_task_invocation_parameters(
self,
input: Option<MaintenanceWindowTaskInvocationParameters>
) -> Self

pub fn priority(self, input: i32) -> Self

The priority of the task in the maintenance window, the lower the number the higher the priority. Tasks in a maintenance window are scheduled in priority order with tasks that have the same priority scheduled in parallel.

pub fn set_priority(self, input: Option<i32>) -> Self

pub fn max_concurrency(self, input: impl Into<String>) -> Self

The maximum number of targets this task can be run for in parallel.

For maintenance window tasks without a target specified, you cannot supply a value for this option. Instead, the system inserts a placeholder value of 1. This value does not affect the running of your task.

pub fn set_max_concurrency(self, input: Option<String>) -> Self

pub fn max_errors(self, input: impl Into<String>) -> Self

The maximum number of errors allowed before this task stops being scheduled.

For maintenance window tasks without a target specified, you cannot supply a value for this option. Instead, the system inserts a placeholder value of 1. This value does not affect the running of your task.

pub fn set_max_errors(self, input: Option<String>) -> Self

pub fn logging_info(self, input: LoggingInfo) -> Self

A structure containing information about an S3 bucket to write instance-level logs to.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

pub fn set_logging_info(self, input: Option<LoggingInfo>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

An optional name for the task.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

An optional description for the task.

pub fn set_description(self, input: Option<String>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

User-provided idempotency token.

pub fn set_client_token(self, input: Option<String>) -> Self