Struct aws_sdk_ssm::client::fluent_builders::SendCommand

pub struct SendCommand<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> SendCommand<C>

pub async fn send(self) -> Result<SendCommandOutput, SdkError<SendCommandError>> where
C: SmithyConnector,

pub fn instance_ids(self, inp: impl Into<String>) -> Self

The IDs of the instances where the command should run. Specifying instance IDs is most useful when you are targeting a limited number of instances, though you can specify up to 50 IDs.

To target a larger number of instances, or if you prefer not to list individual instance IDs, we recommend using the Targets option instead. Using Targets, which accepts tag key-value pairs to identify the instances to send commands to, you can a send command to tens, hundreds, or thousands of instances at once.

For more information about how to use targets, see Using targets and rate controls to send commands to a fleet in the AWS Systems Manager User Guide.

pub fn set_instance_ids(self, input: Option<Vec<String>>) -> Self

pub fn targets(self, inp: impl Into<Target>) -> Self

An array of search criteria that targets instances using a Key,Value combination that you specify. Specifying targets is most useful when you want to send a command to a large number of instances at once. Using Targets, which accepts tag key-value pairs to identify instances, you can send a command to tens, hundreds, or thousands of instances at once.

To send a command to a smaller number of instances, you can use the InstanceIds option instead.

For more information about how to use targets, see Sending commands to a fleet in the AWS Systems Manager User Guide.

pub fn set_targets(self, input: Option<Vec<Target>>) -> Self

pub fn document_name(self, input: impl Into<String>) -> Self

The name of the Systems Manager document to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document ARN. For more information about how to use shared documents, see Using shared SSM documents in the AWS Systems Manager User Guide.

pub fn set_document_name(self, input: Option<String>) -> Self

pub fn document_version(self, input: impl Into<String>) -> Self

The SSM document version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the AWS CLI, then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

--document-version "\$DEFAULT"

--document-version "\$LATEST"

--document-version "3"

pub fn set_document_version(self, input: Option<String>) -> Self

pub fn document_hash(self, input: impl Into<String>) -> Self

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

pub fn set_document_hash(self, input: Option<String>) -> Self

pub fn document_hash_type(self, input: DocumentHashType) -> Self

Sha256 or Sha1.

Sha1 hashes have been deprecated.

pub fn set_document_hash_type(self, input: Option<DocumentHashType>) -> Self

pub fn timeout_seconds(self, input: i32) -> Self

If this time is reached and the command has not already started running, it will not run.

pub fn set_timeout_seconds(self, input: Option<i32>) -> Self

pub fn comment(self, input: impl Into<String>) -> Self

User-specified information about the command, such as a brief description of what the command should do.

pub fn set_comment(self, input: Option<String>) -> Self

pub fn parameters(self, k: impl Into<String>, v: impl Into<Vec<String>>) -> Self

The required and optional parameters specified in the document being run.

pub fn set_parameters(self, input: Option<HashMap<String, Vec<String>>>) -> Self

pub fn output_s3_region(self, input: impl Into<String>) -> Self

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Region of the S3 bucket.

pub fn set_output_s3_region(self, input: Option<String>) -> Self

pub fn output_s3_bucket_name(self, input: impl Into<String>) -> Self

The name of the S3 bucket where command execution responses should be stored.

pub fn set_output_s3_bucket_name(self, input: Option<String>) -> Self

pub fn output_s3_key_prefix(self, input: impl Into<String>) -> Self

The directory structure within the S3 bucket where the responses should be stored.

pub fn set_output_s3_key_prefix(self, input: Option<String>) -> Self

pub fn max_concurrency(self, input: impl Into<String>) -> Self

(Optional) The maximum number of instances that are allowed to run the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Using concurrency controls in the AWS Systems Manager User Guide.

pub fn set_max_concurrency(self, input: Option<String>) -> Self

pub fn max_errors(self, input: impl Into<String>) -> Self

The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 0. For more information about how to use MaxErrors, see Using error controls in the AWS Systems Manager User Guide.

pub fn set_max_errors(self, input: Option<String>) -> Self

pub fn service_role_arn(self, input: impl Into<String>) -> Self

The ARN of the IAM service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for Run Command commands.

pub fn set_service_role_arn(self, input: Option<String>) -> Self

pub fn notification_config(self, input: NotificationConfig) -> Self

Configurations for sending notifications.

pub fn set_notification_config(self, input: Option<NotificationConfig>) -> Self

pub fn cloud_watch_output_config(self, input: CloudWatchOutputConfig) -> Self

Enables Systems Manager to send Run Command output to Amazon CloudWatch Logs.

pub fn set_cloud_watch_output_config(
self,
input: Option<CloudWatchOutputConfig>
) -> Self