Module aws_sdk_ssm::error

Modules

already_exists_exception

See AlreadyExistsException

associated_instances

See AssociatedInstances

association_already_exists

See AssociationAlreadyExists

association_does_not_exist

See AssociationDoesNotExist

association_execution_does_not_exist

See AssociationExecutionDoesNotExist

association_limit_exceeded

See AssociationLimitExceeded

association_version_limit_exceeded

See AssociationVersionLimitExceeded

automation_definition_not_approved_exception

See AutomationDefinitionNotApprovedException

automation_definition_not_found_exception

See AutomationDefinitionNotFoundException

automation_definition_version_not_found_exception

See AutomationDefinitionVersionNotFoundException

automation_execution_limit_exceeded_exception

See AutomationExecutionLimitExceededException

automation_execution_not_found_exception

See AutomationExecutionNotFoundException

automation_step_not_found_exception

See AutomationStepNotFoundException

compliance_type_count_limit_exceeded_exception

See ComplianceTypeCountLimitExceededException

custom_schema_count_limit_exceeded_exception

See CustomSchemaCountLimitExceededException

document_already_exists

See DocumentAlreadyExists

document_limit_exceeded

See DocumentLimitExceeded

document_permission_limit

See DocumentPermissionLimit

document_version_limit_exceeded

See DocumentVersionLimitExceeded

does_not_exist_exception

See DoesNotExistException

duplicate_document_content

See DuplicateDocumentContent

duplicate_document_version_name

See DuplicateDocumentVersionName

duplicate_instance_id

See DuplicateInstanceId

feature_not_available_exception

See FeatureNotAvailableException

hierarchy_level_limit_exceeded_exception

See HierarchyLevelLimitExceededException

hierarchy_type_mismatch_exception

See HierarchyTypeMismatchException

idempotent_parameter_mismatch

See IdempotentParameterMismatch

incompatible_policy_exception

See IncompatiblePolicyException

internal_server_error

See InternalServerError

invalid_activation

See InvalidActivation

invalid_activation_id

See InvalidActivationId

invalid_aggregator_exception

See InvalidAggregatorException

invalid_allowed_pattern_exception

See InvalidAllowedPatternException

invalid_association

See InvalidAssociation

invalid_association_version

See InvalidAssociationVersion

invalid_automation_execution_parameters_exception

See InvalidAutomationExecutionParametersException

invalid_automation_signal_exception

See InvalidAutomationSignalException

invalid_automation_status_update_exception

See InvalidAutomationStatusUpdateException

invalid_command_id

See InvalidCommandId

invalid_delete_inventory_parameters_exception

See InvalidDeleteInventoryParametersException

invalid_deletion_id_exception

See InvalidDeletionIdException

invalid_document

See InvalidDocument

invalid_document_content

See InvalidDocumentContent

invalid_document_operation

See InvalidDocumentOperation

invalid_document_schema_version

See InvalidDocumentSchemaVersion

invalid_document_type

See InvalidDocumentType

invalid_document_version

See InvalidDocumentVersion

invalid_filter

See InvalidFilter

invalid_filter_key

See InvalidFilterKey

invalid_filter_option

See InvalidFilterOption

invalid_filter_value

See InvalidFilterValue

invalid_instance_id

See InvalidInstanceId

invalid_instance_information_filter_value

See InvalidInstanceInformationFilterValue

invalid_inventory_group_exception

See InvalidInventoryGroupException

invalid_inventory_item_context_exception

See InvalidInventoryItemContextException

invalid_inventory_request_exception

See InvalidInventoryRequestException

invalid_item_content_exception

See InvalidItemContentException

invalid_key_id

See InvalidKeyId

invalid_next_token

See InvalidNextToken

invalid_notification_config

See InvalidNotificationConfig

invalid_option_exception

See InvalidOptionException

invalid_output_folder

See InvalidOutputFolder

invalid_output_location

See InvalidOutputLocation

invalid_parameters

See InvalidParameters

invalid_permission_type

See InvalidPermissionType

invalid_plugin_name

See InvalidPluginName

invalid_policy_attribute_exception

See InvalidPolicyAttributeException

invalid_policy_type_exception

See InvalidPolicyTypeException

invalid_resource_id

See InvalidResourceId

invalid_resource_type

See InvalidResourceType

invalid_result_attribute_exception

See InvalidResultAttributeException

invalid_role

See InvalidRole

invalid_schedule

See InvalidSchedule

invalid_target

See InvalidTarget

invalid_type_name_exception

See InvalidTypeNameException

invalid_update

See InvalidUpdate

invocation_does_not_exist

See InvocationDoesNotExist

item_content_mismatch_exception

See ItemContentMismatchException

item_size_limit_exceeded_exception

See ItemSizeLimitExceededException

max_document_size_exceeded

See MaxDocumentSizeExceeded

ops_item_already_exists_exception

See OpsItemAlreadyExistsException

ops_item_invalid_parameter_exception

See OpsItemInvalidParameterException

ops_item_limit_exceeded_exception

See OpsItemLimitExceededException

ops_item_not_found_exception

See OpsItemNotFoundException

ops_item_related_item_already_exists_exception

See OpsItemRelatedItemAlreadyExistsException

ops_item_related_item_association_not_found_exception

See OpsItemRelatedItemAssociationNotFoundException

ops_metadata_already_exists_exception

See OpsMetadataAlreadyExistsException

ops_metadata_invalid_argument_exception

See OpsMetadataInvalidArgumentException

ops_metadata_key_limit_exceeded_exception

See OpsMetadataKeyLimitExceededException

ops_metadata_limit_exceeded_exception

See OpsMetadataLimitExceededException

ops_metadata_not_found_exception

See OpsMetadataNotFoundException

ops_metadata_too_many_updates_exception

See OpsMetadataTooManyUpdatesException

parameter_already_exists

See ParameterAlreadyExists

parameter_limit_exceeded

See ParameterLimitExceeded

parameter_max_version_limit_exceeded

See ParameterMaxVersionLimitExceeded

parameter_not_found

See ParameterNotFound

parameter_pattern_mismatch_exception

See ParameterPatternMismatchException

parameter_version_label_limit_exceeded

See ParameterVersionLabelLimitExceeded

parameter_version_not_found

See ParameterVersionNotFound

policies_limit_exceeded_exception

See PoliciesLimitExceededException

resource_data_sync_already_exists_exception

See ResourceDataSyncAlreadyExistsException

resource_data_sync_conflict_exception

See ResourceDataSyncConflictException

resource_data_sync_count_exceeded_exception

See ResourceDataSyncCountExceededException

resource_data_sync_invalid_configuration_exception

See ResourceDataSyncInvalidConfigurationException

resource_data_sync_not_found_exception

See ResourceDataSyncNotFoundException

resource_in_use_exception

See ResourceInUseException

resource_limit_exceeded_exception

See ResourceLimitExceededException

service_setting_not_found

See ServiceSettingNotFound

status_unchanged

See StatusUnchanged

sub_type_count_limit_exceeded_exception

See SubTypeCountLimitExceededException

target_in_use_exception

See TargetInUseException

target_not_connected

See TargetNotConnected

too_many_tags_error

See TooManyTagsError

too_many_updates

See TooManyUpdates

total_size_limit_exceeded_exception

See TotalSizeLimitExceededException

unsupported_calendar_exception

See UnsupportedCalendarException

unsupported_feature_required_exception

See UnsupportedFeatureRequiredException

unsupported_inventory_item_context_exception

See UnsupportedInventoryItemContextException

unsupported_inventory_schema_version_exception

See UnsupportedInventorySchemaVersionException

unsupported_operating_system

See UnsupportedOperatingSystem

unsupported_parameter_type

See UnsupportedParameterType

unsupported_platform_type

See UnsupportedPlatformType

Structs

AddTagsToResourceError

AlreadyExistsException

Error returned if an attempt is made to register a patch group with a patch baseline that is

AssociateOpsItemRelatedItemError

AssociatedInstances

You must disassociate a document from all instances before you can delete it.

AssociationAlreadyExists

The specified association already exists.

AssociationDoesNotExist

The specified association does not exist.

AssociationExecutionDoesNotExist

The specified execution ID does not exist. Verify the ID number and try again.

AssociationLimitExceeded

You can have at most 2,000 active associations.

AssociationVersionLimitExceeded

You have reached the maximum number versions allowed for an association. Each association

AutomationDefinitionNotApprovedException

Indicates that the Change Manager change template used in the change request was rejected or is

AutomationDefinitionNotFoundException

An Automation document with the specified name could not be found.

AutomationDefinitionVersionNotFoundException

An Automation document with the specified name and version could not be found.

AutomationExecutionLimitExceededException

The number of simultaneously running Automation executions exceeded the allowable

AutomationExecutionNotFoundException

There is no automation execution information for the requested automation execution

AutomationStepNotFoundException

The specified step name and execution ID don't exist. Verify the information and try

CancelCommandError

CancelMaintenanceWindowExecutionError

ComplianceTypeCountLimitExceededException

You specified too many custom compliance types. You can specify a maximum of 10 different

CreateActivationError

CreateAssociationBatchError

CreateAssociationError

CreateDocumentError

CreateMaintenanceWindowError

CreateOpsItemError

CreateOpsMetadataError

CreatePatchBaselineError

CreateResourceDataSyncError

CustomSchemaCountLimitExceededException

You have exceeded the limit for custom schemas. Delete one or more custom schemas and try

DeleteActivationError

DeleteAssociationError

DeleteDocumentError

DeleteInventoryError

DeleteMaintenanceWindowError

DeleteOpsMetadataError

DeleteParameterError

DeleteParametersError

DeletePatchBaselineError

DeleteResourceDataSyncError

DeregisterManagedInstanceError

DeregisterPatchBaselineForPatchGroupError

DeregisterTargetFromMaintenanceWindowError

DeregisterTaskFromMaintenanceWindowError

DescribeActivationsError

DescribeAssociationError

DescribeAssociationExecutionTargetsError

DescribeAssociationExecutionsError

DescribeAutomationExecutionsError

DescribeAutomationStepExecutionsError

DescribeAvailablePatchesError

DescribeDocumentError

DescribeDocumentPermissionError

DescribeEffectiveInstanceAssociationsError

DescribeEffectivePatchesForPatchBaselineError

DescribeInstanceAssociationsStatusError

DescribeInstanceInformationError

DescribeInstancePatchStatesError

DescribeInstancePatchStatesForPatchGroupError

DescribeInstancePatchesError

DescribeInventoryDeletionsError

DescribeMaintenanceWindowExecutionTaskInvocationsError

DescribeMaintenanceWindowExecutionTasksError

DescribeMaintenanceWindowExecutionsError

DescribeMaintenanceWindowScheduleError

DescribeMaintenanceWindowTargetsError

DescribeMaintenanceWindowTasksError

DescribeMaintenanceWindowsError

DescribeMaintenanceWindowsForTargetError

DescribeOpsItemsError

DescribeParametersError

DescribePatchBaselinesError

DescribePatchGroupStateError

DescribePatchGroupsError

DescribePatchPropertiesError

DescribeSessionsError

DisassociateOpsItemRelatedItemError

DocumentAlreadyExists

The specified document already exists.

DocumentLimitExceeded

You can have at most 500 active Systems Manager documents.

DocumentPermissionLimit

The document cannot be shared with more AWS user accounts. You can share a document with a

DocumentVersionLimitExceeded

The document has too many versions. Delete one or more document versions and try

DoesNotExistException

Error returned when the ID specified for a resource, such as a maintenance window or Patch

DuplicateDocumentContent

The content of the association document matches another document. Change the content of the

DuplicateDocumentVersionName

The version name has already been used in this document. Specify a different version name,

DuplicateInstanceId

You cannot specify an instance ID in more than one association.

FeatureNotAvailableException

You attempted to register a LAMBDA or STEP_FUNCTIONS task in a region where the

GetAutomationExecutionError

GetCalendarStateError

GetCommandInvocationError

GetConnectionStatusError

GetDefaultPatchBaselineError

GetDeployablePatchSnapshotForInstanceError

GetDocumentError

GetInventoryError

GetInventorySchemaError

GetMaintenanceWindowError

GetMaintenanceWindowExecutionError

GetMaintenanceWindowExecutionTaskError

GetMaintenanceWindowExecutionTaskInvocationError

GetMaintenanceWindowTaskError

GetOpsItemError

GetOpsMetadataError

GetOpsSummaryError

GetParameterError

GetParameterHistoryError

GetParametersByPathError

GetParametersError

GetPatchBaselineError

GetPatchBaselineForPatchGroupError

GetServiceSettingError

HierarchyLevelLimitExceededException

A hierarchy can have a maximum of 15 levels. For more information, see Requirements and

HierarchyTypeMismatchException

Parameter Store does not support changing a parameter type in a hierarchy. For example, you

IdempotentParameterMismatch

Error returned when an idempotent operation is retried and the parameters don't match the

IncompatiblePolicyException

There is a conflict in the policies specified for this parameter. You can't, for example,

InternalServerError

An error occurred on the server side.

InvalidActivation

The activation is not valid. The activation might have been deleted, or the ActivationId and

InvalidActivationId

The activation ID is not valid. Verify the you entered the correct ActivationId or

InvalidAggregatorException

The specified aggregator is not valid for inventory groups. Verify that the aggregator uses

InvalidAllowedPatternException

The request does not meet the regular expression requirement.

InvalidAssociation

The association is not valid or does not exist.

InvalidAssociationVersion

The version you specified is not valid. Use ListAssociationVersions to view all versions of

InvalidAutomationExecutionParametersException

The supplied parameters for invoking the specified Automation document are incorrect. For

InvalidAutomationSignalException

The signal is not valid for the current Automation execution.

InvalidAutomationStatusUpdateException

The specified update status operation is not valid.

InvalidCommandId

The specified command ID is not valid. Verify the ID and try again.

InvalidDeleteInventoryParametersException

One or more of the parameters specified for the delete operation is not valid. Verify all

InvalidDeletionIdException

The ID specified for the delete operation does not exist or is not valid. Verify the ID and

InvalidDocument

The specified document does not exist.

InvalidDocumentContent

The content for the document is not valid.

InvalidDocumentOperation

You attempted to delete a document while it is still shared. You must stop sharing the

InvalidDocumentSchemaVersion

The version of the document schema is not supported.

InvalidDocumentType

The document type is not valid. Valid document types are described in the

InvalidDocumentVersion

The document version is not valid or does not exist.

InvalidFilter

The filter name is not valid. Verify the you entered the correct name and try again.

InvalidFilterKey

The specified key is not valid.

InvalidFilterOption

The specified filter option is not valid. Valid options are Equals and BeginsWith. For Path

InvalidFilterValue

The filter value is not valid. Verify the value and try again.

InvalidInstanceId

The following problems can cause this exception:

InvalidInstanceInformationFilterValue

The specified filter value is not valid.

InvalidInventoryGroupException

The specified inventory group is not valid.

InvalidInventoryItemContextException

You specified invalid keys or values in the Context attribute for

InvalidInventoryRequestException

The request is not valid.

InvalidItemContentException

One or more content items is not valid.

InvalidKeyId

The query key ID is not valid.

InvalidNextToken

The specified token is not valid.

InvalidNotificationConfig

One or more configuration items is not valid. Verify that a valid Amazon Resource Name (ARN)

InvalidOptionException

The delete inventory option specified is not valid. Verify the option and try again.

InvalidOutputFolder

The S3 bucket does not exist.

InvalidOutputLocation

The output location is not valid or does not exist.

InvalidParameters

You must specify values for all required parameters in the Systems Manager document. You can only

InvalidPermissionType

The permission type is not supported. Share is the only supported

InvalidPluginName

The plugin name is not valid.

InvalidPolicyAttributeException

A policy attribute or its value is invalid.

InvalidPolicyTypeException

The policy type is not supported. Parameter Store supports the following policy types:

InvalidResourceId

The resource ID is not valid. Verify that you entered the correct ID and try again.

InvalidResourceType

The resource type is not valid. For example, if you are attempting to tag an instance, the

InvalidResultAttributeException

The specified inventory item result attribute is not valid.

InvalidRole

The role name can't contain invalid characters. Also verify that you specified an IAM role

InvalidSchedule

The schedule is invalid. Verify your cron or rate expression and try again.

InvalidTarget

The target is not valid or does not exist. It might not be configured for Systems Manager or you might

InvalidTypeNameException

The parameter type name is not valid.

InvalidUpdate

The update is not valid.

InvocationDoesNotExist

The command ID and instance ID you specified did not match any invocations. Verify the

ItemContentMismatchException

The inventory item has invalid content.

ItemSizeLimitExceededException

The inventory item size has exceeded the size limit.

LabelParameterVersionError

ListAssociationVersionsError

ListAssociationsError

ListCommandInvocationsError

ListCommandsError

ListComplianceItemsError

ListComplianceSummariesError

ListDocumentMetadataHistoryError

ListDocumentVersionsError

ListDocumentsError

ListInventoryEntriesError

ListOpsItemEventsError

ListOpsItemRelatedItemsError

ListOpsMetadataError

ListResourceComplianceSummariesError

ListResourceDataSyncError

ListTagsForResourceError

MaxDocumentSizeExceeded

The size limit of a document is 64 KB.

ModifyDocumentPermissionError

OpsItemAlreadyExistsException

The OpsItem already exists.

OpsItemInvalidParameterException

A specified parameter argument isn't valid. Verify the available arguments and try

OpsItemLimitExceededException

The request caused OpsItems to exceed one or more quotas. For information about OpsItem

OpsItemNotFoundException

The specified OpsItem ID doesn't exist. Verify the ID and try again.

OpsItemRelatedItemAlreadyExistsException

The Amazon Resource Name (ARN) is already associated with the OpsItem.

OpsItemRelatedItemAssociationNotFoundException

The association was not found using the parameters you specified in the call. Verify the

OpsMetadataAlreadyExistsException

An OpsMetadata object already exists for the selected resource.

OpsMetadataInvalidArgumentException

One of the arguments passed is invalid.

OpsMetadataKeyLimitExceededException

The OpsMetadata object exceeds the maximum number of OpsMetadata keys that you can assign to

OpsMetadataLimitExceededException

Your account reached the maximum number of OpsMetadata objects allowed by Application Manager. The

OpsMetadataNotFoundException

The OpsMetadata object does not exist.

OpsMetadataTooManyUpdatesException

The system is processing too many concurrent updates. Wait a few moments and try

ParameterAlreadyExists

The parameter already exists. You can't create duplicate parameters.

ParameterLimitExceeded

You have exceeded the number of parameters for this AWS account. Delete one or more

ParameterMaxVersionLimitExceeded

Parameter Store retains the 100 most recently created versions of a parameter. After this

ParameterNotFound

The parameter could not be found. Verify the name and try again.

ParameterPatternMismatchException

The parameter name is not valid.

ParameterVersionLabelLimitExceeded

A parameter version can have a maximum of ten labels.

ParameterVersionNotFound

The specified parameter version was not found. Verify the parameter name and version, and

PoliciesLimitExceededException

You specified more than the maximum number of allowed policies for the parameter. The

PutComplianceItemsError

PutInventoryError

PutParameterError

RegisterDefaultPatchBaselineError

RegisterPatchBaselineForPatchGroupError

RegisterTargetWithMaintenanceWindowError

RegisterTaskWithMaintenanceWindowError

RemoveTagsFromResourceError

ResetServiceSettingError

ResourceDataSyncAlreadyExistsException

A sync configuration with the same name already exists.

ResourceDataSyncConflictException

Another UpdateResourceDataSync request is being processed. Wait a few minutes

ResourceDataSyncCountExceededException

You have exceeded the allowed maximum sync configurations.

ResourceDataSyncInvalidConfigurationException

The specified sync configuration is invalid.

ResourceDataSyncNotFoundException

The specified sync name was not found.

ResourceInUseException

Error returned if an attempt is made to delete a patch baseline that is registered for a

ResourceLimitExceededException

Error returned when the caller has exceeded the default resource quotas. For example, too

ResumeSessionError

SendAutomationSignalError

SendCommandError

ServiceSettingNotFound

The specified service setting was not found. Either the service name or the setting has not

StartAssociationsOnceError

StartAutomationExecutionError

StartChangeRequestExecutionError

StartSessionError

StatusUnchanged

The updated status is the same as the current status.

StopAutomationExecutionError

SubTypeCountLimitExceededException

The sub-type count exceeded the limit for the inventory type.

TargetInUseException

You specified the Safe option for the DeregisterTargetFromMaintenanceWindow

TargetNotConnected

The specified target instance for the session is not fully configured for use with Session Manager.

TerminateSessionError

TooManyTagsError

The Targets parameter includes too many tags. Remove one or more tags and try

TooManyUpdates

There are concurrent updates for a resource that supports one update at a time.

TotalSizeLimitExceededException

The size of inventory data has exceeded the total size limit for the resource.

UnlabelParameterVersionError

UnsupportedCalendarException

The calendar entry contained in the specified Systems Manager document is not supported.

UnsupportedFeatureRequiredException

Microsoft application patching is only available on EC2 instances and advanced instances. To

UnsupportedInventoryItemContextException

The Context attribute that you specified for the InventoryItem is

UnsupportedInventorySchemaVersionException

Inventory item type schema version has to match supported versions in the service. Check

UnsupportedOperatingSystem

The operating systems you specified is not supported, or the operation is not supported for

UnsupportedParameterType

The parameter type is not supported.

UnsupportedPlatformType

The document does not support the platform type of the given instance ID(s). For example,

UpdateAssociationError

UpdateAssociationStatusError

UpdateDocumentDefaultVersionError

UpdateDocumentError

UpdateDocumentMetadataError

UpdateMaintenanceWindowError

UpdateMaintenanceWindowTargetError

UpdateMaintenanceWindowTaskError

UpdateManagedInstanceRoleError

UpdateOpsItemError

UpdateOpsMetadataError

UpdatePatchBaselineError

UpdateResourceDataSyncError

UpdateServiceSettingError

Enums

AddTagsToResourceErrorKind

AssociateOpsItemRelatedItemErrorKind

CancelCommandErrorKind

CancelMaintenanceWindowExecutionErrorKind

CreateActivationErrorKind

CreateAssociationBatchErrorKind

CreateAssociationErrorKind

CreateDocumentErrorKind

CreateMaintenanceWindowErrorKind

CreateOpsItemErrorKind

CreateOpsMetadataErrorKind

CreatePatchBaselineErrorKind

CreateResourceDataSyncErrorKind

DeleteActivationErrorKind

DeleteAssociationErrorKind

DeleteDocumentErrorKind

DeleteInventoryErrorKind

DeleteMaintenanceWindowErrorKind

DeleteOpsMetadataErrorKind

DeleteParameterErrorKind

DeleteParametersErrorKind

DeletePatchBaselineErrorKind

DeleteResourceDataSyncErrorKind

DeregisterManagedInstanceErrorKind

DeregisterPatchBaselineForPatchGroupErrorKind

DeregisterTargetFromMaintenanceWindowErrorKind

DeregisterTaskFromMaintenanceWindowErrorKind

DescribeActivationsErrorKind

DescribeAssociationErrorKind

DescribeAssociationExecutionTargetsErrorKind

DescribeAssociationExecutionsErrorKind

DescribeAutomationExecutionsErrorKind

DescribeAutomationStepExecutionsErrorKind

DescribeAvailablePatchesErrorKind

DescribeDocumentErrorKind

DescribeDocumentPermissionErrorKind

DescribeEffectiveInstanceAssociationsErrorKind

DescribeEffectivePatchesForPatchBaselineErrorKind

DescribeInstanceAssociationsStatusErrorKind

DescribeInstanceInformationErrorKind

DescribeInstancePatchStatesErrorKind

DescribeInstancePatchStatesForPatchGroupErrorKind

DescribeInstancePatchesErrorKind

DescribeInventoryDeletionsErrorKind

DescribeMaintenanceWindowExecutionTaskInvocationsErrorKind

DescribeMaintenanceWindowExecutionTasksErrorKind

DescribeMaintenanceWindowExecutionsErrorKind

DescribeMaintenanceWindowScheduleErrorKind

DescribeMaintenanceWindowTargetsErrorKind

DescribeMaintenanceWindowTasksErrorKind

DescribeMaintenanceWindowsErrorKind

DescribeMaintenanceWindowsForTargetErrorKind

DescribeOpsItemsErrorKind

DescribeParametersErrorKind

DescribePatchBaselinesErrorKind

DescribePatchGroupStateErrorKind

DescribePatchGroupsErrorKind

DescribePatchPropertiesErrorKind

DescribeSessionsErrorKind

DisassociateOpsItemRelatedItemErrorKind

GetAutomationExecutionErrorKind

GetCalendarStateErrorKind

GetCommandInvocationErrorKind

GetConnectionStatusErrorKind

GetDefaultPatchBaselineErrorKind

GetDeployablePatchSnapshotForInstanceErrorKind

GetDocumentErrorKind

GetInventoryErrorKind

GetInventorySchemaErrorKind

GetMaintenanceWindowErrorKind

GetMaintenanceWindowExecutionErrorKind

GetMaintenanceWindowExecutionTaskErrorKind

GetMaintenanceWindowExecutionTaskInvocationErrorKind

GetMaintenanceWindowTaskErrorKind

GetOpsItemErrorKind

GetOpsMetadataErrorKind

GetOpsSummaryErrorKind

GetParameterErrorKind

GetParameterHistoryErrorKind

GetParametersByPathErrorKind

GetParametersErrorKind

GetPatchBaselineErrorKind

GetPatchBaselineForPatchGroupErrorKind

GetServiceSettingErrorKind

LabelParameterVersionErrorKind

ListAssociationVersionsErrorKind

ListAssociationsErrorKind

ListCommandInvocationsErrorKind

ListCommandsErrorKind

ListComplianceItemsErrorKind

ListComplianceSummariesErrorKind

ListDocumentMetadataHistoryErrorKind

ListDocumentVersionsErrorKind

ListDocumentsErrorKind

ListInventoryEntriesErrorKind

ListOpsItemEventsErrorKind

ListOpsItemRelatedItemsErrorKind

ListOpsMetadataErrorKind

ListResourceComplianceSummariesErrorKind

ListResourceDataSyncErrorKind

ListTagsForResourceErrorKind

ModifyDocumentPermissionErrorKind

PutComplianceItemsErrorKind

PutInventoryErrorKind

PutParameterErrorKind

RegisterDefaultPatchBaselineErrorKind

RegisterPatchBaselineForPatchGroupErrorKind

RegisterTargetWithMaintenanceWindowErrorKind

RegisterTaskWithMaintenanceWindowErrorKind

RemoveTagsFromResourceErrorKind

ResetServiceSettingErrorKind

ResumeSessionErrorKind

SendAutomationSignalErrorKind

SendCommandErrorKind

StartAssociationsOnceErrorKind

StartAutomationExecutionErrorKind

StartChangeRequestExecutionErrorKind

StartSessionErrorKind

StopAutomationExecutionErrorKind

TerminateSessionErrorKind

UnlabelParameterVersionErrorKind

UpdateAssociationErrorKind

UpdateAssociationStatusErrorKind

UpdateDocumentDefaultVersionErrorKind

UpdateDocumentErrorKind

UpdateDocumentMetadataErrorKind

UpdateMaintenanceWindowErrorKind

UpdateMaintenanceWindowTargetErrorKind

UpdateMaintenanceWindowTaskErrorKind

UpdateManagedInstanceRoleErrorKind

UpdateOpsItemErrorKind

UpdateOpsMetadataErrorKind

UpdatePatchBaselineErrorKind

UpdateResourceDataSyncErrorKind

UpdateServiceSettingErrorKind