Struct aws_sdk_ssm::input::CreateAssociationInput

#[non_exhaustive]
pub struct CreateAssociationInput {

    pub name: Option<String>,
    pub document_version: Option<String>,
    pub instance_id: Option<String>,
    pub parameters: Option<HashMap<String, Vec<String>>>,
    pub targets: Option<Vec<Target>>,
    pub schedule_expression: Option<String>,
    pub output_location: Option<InstanceAssociationOutputLocation>,
    pub association_name: Option<String>,
    pub automation_target_parameter_name: Option<String>,
    pub max_errors: Option<String>,
    pub max_concurrency: Option<String>,
    pub compliance_severity: Option<AssociationComplianceSeverity>,
    pub sync_compliance: Option<AssociationSyncCompliance>,
    pub apply_only_at_cron_interval: bool,
    pub calendar_names: Option<Vec<String>>,
    pub target_locations: Option<Vec<TargetLocation>>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The name of the SSM document that contains the configuration information for the instance. You can specify Command or Automation documents.

You can specify AWS-predefined documents, documents you created, or a document that is shared with you from another account.

For SSM documents that are shared with you from other AWS accounts, you must specify the complete SSM document ARN, in the following format:

arn:partition:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For AWS-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

document_version: Option<String>

The document version you want to associate with the target(s). Can be a specific version or the default version.

instance_id: Option<String>

The instance ID.

InstanceId has been deprecated. To specify an instance ID for an association, use the Targets parameter. Requests that include the parameter InstanceID with SSM documents that use schema version 2.0 or later will fail. In addition, if you use the parameter InstanceId, you cannot use the parameters AssociationName, DocumentVersion, MaxErrors, MaxConcurrency, OutputLocation, or ScheduleExpression. To use these parameters, you must use the Targets parameter.

parameters: Option<HashMap<String, Vec<String>>>

The parameters for the runtime configuration of the document.

targets: Option<Vec<Target>>

The targets for the association. You can target instances by using tags, AWS Resource Groups, all instances in an AWS account, or individual instance IDs. For more information about choosing targets for an association, see Using targets and rate controls with State Manager associations in the AWS Systems Manager User Guide.

schedule_expression: Option<String>

A cron expression when the association will be applied to the target(s).

output_location: Option<InstanceAssociationOutputLocation>

An S3 bucket where you want to store the output details of the request.

association_name: Option<String>

Specify a descriptive name for the association.

automation_target_parameter_name: Option<String>

Specify the target for the association. This target is required for associations that use an Automation document and target resources by using rate controls.

max_errors: Option<String>

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

max_concurrency: Option<String>

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

compliance_severity: Option<AssociationComplianceSeverity>

The severity level to assign to the association.

sync_compliance: Option<AssociationSyncCompliance>

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API action. In this case, compliance data is not managed by State Manager. It is managed by your direct call to the PutComplianceItems API action.

By default, all associations use AUTO mode.

apply_only_at_cron_interval: bool

By default, when you create a new association, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter is not supported for rate expressions.

calendar_names: Option<Vec<String>>

The names or Amazon Resource Names (ARNs) of the Systems Manager Change Calendar type documents you want to gate your associations under. The associations only run when that Change Calendar is open. For more information, see AWS Systems Manager Change Calendar.

target_locations: Option<Vec<TargetLocation>>

A location is a combination of AWS Regions and AWS accounts where you want to run the association. Use this action to create an association in multiple Regions and multiple accounts.

Implementations

impl CreateAssociationInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<CreateAssociation, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateAssociation>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateAssociationInput