Struct aws_sdk_ssm::input::CreateOpsItemInput

#[non_exhaustive]
pub struct CreateOpsItemInput {

    pub description: Option<String>,
    pub ops_item_type: Option<String>,
    pub operational_data: Option<HashMap<String, OpsItemDataValue>>,
    pub notifications: Option<Vec<OpsItemNotification>>,
    pub priority: Option<i32>,
    pub related_ops_items: Option<Vec<RelatedOpsItem>>,
    pub source: Option<String>,
    pub title: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub category: Option<String>,
    pub severity: Option<String>,
    pub actual_start_time: Option<Instant>,
    pub actual_end_time: Option<Instant>,
    pub planned_start_time: Option<Instant>,
    pub planned_end_time: Option<Instant>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

description: Option<String>

Information about the OpsItem.

ops_item_type: Option<String>

The type of OpsItem to create. Currently, the only valid values are /aws/changerequest and /aws/issue.

operational_data: Option<HashMap<String, OpsItemDataValue>>

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API action) can view and search on the specified data. Operational data that is not searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API action).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view AWS CLI example commands that use these keys, see Creating OpsItems manually in the AWS Systems Manager User Guide.

notifications: Option<Vec<OpsItemNotification>>

The Amazon Resource Name (ARN) of an SNS topic where notifications are sent when this OpsItem is edited or changed.

priority: Option<i32>

The importance of this OpsItem in relation to other OpsItems in the system.

related_ops_items: Option<Vec<RelatedOpsItem>>

One or more OpsItems that share something in common with the current OpsItems. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

source: Option<String>

The origin of the OpsItem, such as Amazon EC2 or Systems Manager.

The source name can't contain the following strings: aws, amazon, and amzn.

title: Option<String>

A short heading that describes the nature of the OpsItem and the impacted resource.

tags: Option<Vec<Tag>>

Optional metadata that you assign to a resource. You can restrict access to OpsItems by using an inline IAM policy that specifies tags. For more information, see Getting started with OpsCenter in the AWS Systems Manager User Guide.

Tags use a key-value pair. For example:

Key=Department,Value=Finance

To add tags to an existing OpsItem, use the AddTagsToResource action.

category: Option<String>

Specify a category to assign to an OpsItem.

severity: Option<String>

Specify a severity to assign to an OpsItem.

actual_start_time: Option<Instant>

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

actual_end_time: Option<Instant>

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

planned_start_time: Option<Instant>

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

planned_end_time: Option<Instant>

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

Implementations

impl CreateOpsItemInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<CreateOpsItem, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateOpsItem>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateOpsItemInput