Struct aws_sdk_ssm::input::SendCommandInput

#[non_exhaustive]
pub struct SendCommandInput {

    pub instance_ids: Option<Vec<String>>,
    pub targets: Option<Vec<Target>>,
    pub document_name: Option<String>,
    pub document_version: Option<String>,
    pub document_hash: Option<String>,
    pub document_hash_type: Option<DocumentHashType>,
    pub timeout_seconds: Option<i32>,
    pub comment: Option<String>,
    pub parameters: Option<HashMap<String, Vec<String>>>,
    pub output_s3_region: Option<String>,
    pub output_s3_bucket_name: Option<String>,
    pub output_s3_key_prefix: Option<String>,
    pub max_concurrency: Option<String>,
    pub max_errors: Option<String>,
    pub service_role_arn: Option<String>,
    pub notification_config: Option<NotificationConfig>,
    pub cloud_watch_output_config: Option<CloudWatchOutputConfig>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

instance_ids: Option<Vec<String>>

The IDs of the instances where the command should run. Specifying instance IDs is most useful when you are targeting a limited number of instances, though you can specify up to 50 IDs.

To target a larger number of instances, or if you prefer not to list individual instance IDs, we recommend using the Targets option instead. Using Targets, which accepts tag key-value pairs to identify the instances to send commands to, you can a send command to tens, hundreds, or thousands of instances at once.

For more information about how to use targets, see Using targets and rate controls to send commands to a fleet in the AWS Systems Manager User Guide.

targets: Option<Vec<Target>>

An array of search criteria that targets instances using a Key,Value combination that you specify. Specifying targets is most useful when you want to send a command to a large number of instances at once. Using Targets, which accepts tag key-value pairs to identify instances, you can send a command to tens, hundreds, or thousands of instances at once.

To send a command to a smaller number of instances, you can use the InstanceIds option instead.

For more information about how to use targets, see Sending commands to a fleet in the AWS Systems Manager User Guide.

document_name: Option<String>

The name of the Systems Manager document to run. This can be a public document or a custom document. To run a shared document belonging to another account, specify the document ARN. For more information about how to use shared documents, see Using shared SSM documents in the AWS Systems Manager User Guide.

document_version: Option<String>

The SSM document version to use in the request. You can specify $DEFAULT, $LATEST, or a specific version number. If you run commands by using the AWS CLI, then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

--document-version "\$DEFAULT"

--document-version "\$LATEST"

--document-version "3"

document_hash: Option<String>

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

document_hash_type: Option<DocumentHashType>

Sha256 or Sha1.

Sha1 hashes have been deprecated.

timeout_seconds: Option<i32>

If this time is reached and the command has not already started running, it will not run.

comment: Option<String>

User-specified information about the command, such as a brief description of what the command should do.

parameters: Option<HashMap<String, Vec<String>>>

The required and optional parameters specified in the document being run.

output_s3_region: Option<String>

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Region of the S3 bucket.

output_s3_bucket_name: Option<String>

The name of the S3 bucket where command execution responses should be stored.

output_s3_key_prefix: Option<String>

The directory structure within the S3 bucket where the responses should be stored.

max_concurrency: Option<String>

(Optional) The maximum number of instances that are allowed to run the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Using concurrency controls in the AWS Systems Manager User Guide.

max_errors: Option<String>

The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 0. For more information about how to use MaxErrors, see Using error controls in the AWS Systems Manager User Guide.

service_role_arn: Option<String>

The ARN of the IAM service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for Run Command commands.

notification_config: Option<NotificationConfig>

Configurations for sending notifications.

cloud_watch_output_config: Option<CloudWatchOutputConfig>

Enables Systems Manager to send Run Command output to Amazon CloudWatch Logs.

Implementations

impl SendCommandInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<SendCommand, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<SendCommand>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture SendCommandInput