Struct aws_sdk_ssm::input::UpdateMaintenanceWindowTaskInput

#[non_exhaustive]
pub struct UpdateMaintenanceWindowTaskInput {

    pub window_id: Option<String>,
    pub window_task_id: Option<String>,
    pub targets: Option<Vec<Target>>,
    pub task_arn: Option<String>,
    pub service_role_arn: Option<String>,
    pub task_parameters: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>,
    pub task_invocation_parameters: Option<MaintenanceWindowTaskInvocationParameters>,
    pub priority: Option<i32>,
    pub max_concurrency: Option<String>,
    pub max_errors: Option<String>,
    pub logging_info: Option<LoggingInfo>,
    pub name: Option<String>,
    pub description: Option<String>,
    pub replace: Option<bool>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

window_id: Option<String>

The maintenance window ID that contains the task to modify.

window_task_id: Option<String>

The task ID to modify.

targets: Option<Vec<Target>>

The targets (either instances or tags) to modify. Instances are specified using Key=instanceids,Values=instanceID_1,instanceID_2. Tags are specified using Key=tag_name,Values=tag_value.

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, AWS Lambda, and AWS Step Functions). For more information about running tasks that do not specify targets, see Registering maintenance window tasks without targets in the AWS Systems Manager User Guide.

task_arn: Option<String>

The task ARN to modify.

service_role_arn: Option<String>

The ARN of the IAM service role for Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the AWS Systems Manager User Guide:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

task_parameters: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>

The parameters to modify.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

The map has the following format:

Key: string, between 1 and 255 characters

Value: an array of strings, each string is between 1 and 255 characters

task_invocation_parameters: Option<MaintenanceWindowTaskInvocationParameters>

The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

When you update a maintenance window task that has options specified in TaskInvocationParameters, you must provide again all the TaskInvocationParameters values that you want to retain. The values you do not specify again are removed. For example, suppose that when you registered a Run Command task, you specified TaskInvocationParameters values for Comment, NotificationConfig, and OutputS3BucketName. If you update the maintenance window task and specify only a different OutputS3BucketName value, the values for Comment and NotificationConfig are removed.

priority: Option<i32>

The new task priority to specify. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

max_concurrency: Option<String>

The new MaxConcurrency value you want to specify. MaxConcurrency is the number of targets that are allowed to run this task in parallel.

For maintenance window tasks without a target specified, you cannot supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value does not affect the running of your task and can be ignored.

max_errors: Option<String>

The new MaxErrors value to specify. MaxErrors is the maximum number of errors that are allowed before the task stops being scheduled.

For maintenance window tasks without a target specified, you cannot supply a value for this option. Instead, the system inserts a placeholder value of 1, which may be reported in the response to this command. This value does not affect the running of your task and can be ignored.

logging_info: Option<LoggingInfo>

The new logging location in Amazon S3 to specify.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

name: Option<String>

The new task name to specify.

description: Option<String>

The new task description to specify.

replace: Option<bool>

If True, then all fields that are required by the RegisterTaskWithMaintenanceWindow action are also required for this API request. Optional fields that are not specified are set to null.

Implementations

impl UpdateMaintenanceWindowTaskInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<UpdateMaintenanceWindowTask, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<UpdateMaintenanceWindowTask>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture UpdateMaintenanceWindowTaskInput