Struct aws_sdk_ssm::input::update_maintenance_window_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UpdateMaintenanceWindowInput

Implementations

impl Builder

pub fn window_id(self, input: impl Into<String>) -> Self

The ID of the maintenance window to update.

pub fn set_window_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name of the maintenance window.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

An optional description for the update request.

pub fn set_description(self, input: Option<String>) -> Self

pub fn start_date(self, input: impl Into<String>) -> Self

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

pub fn set_start_date(self, input: Option<String>) -> Self

pub fn end_date(self, input: impl Into<String>) -> Self

The date and time, in ISO-8601 Extended format, for when you want the maintenance window to become inactive. EndDate allows you to set a date and time in the future when the maintenance window will no longer run.

pub fn set_end_date(self, input: Option<String>) -> Self

pub fn schedule(self, input: impl Into<String>) -> Self

The schedule of the maintenance window in the form of a cron or rate expression.

pub fn set_schedule(self, input: Option<String>) -> Self

pub fn schedule_timezone(self, input: impl Into<String>) -> Self

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

pub fn set_schedule_timezone(self, input: Option<String>) -> Self

pub fn schedule_offset(self, input: i32) -> Self

The number of days to wait after the date and time specified by a CRON expression before running the maintenance window.

For example, the following cron expression schedules a maintenance window to run the third Tuesday of every month at 11:30 PM.

cron(30 23 ? * TUE#3 *)

If the schedule offset is 2, the maintenance window won't run until two days later.

pub fn set_schedule_offset(self, input: Option<i32>) -> Self

pub fn duration(self, input: i32) -> Self

The duration of the maintenance window in hours.

pub fn set_duration(self, input: Option<i32>) -> Self

pub fn cutoff(self, input: i32) -> Self

The number of hours before the end of the maintenance window that Systems Manager stops scheduling new tasks for execution.

pub fn set_cutoff(self, input: Option<i32>) -> Self

pub fn allow_unassociated_targets(self, input: bool) -> Self

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

pub fn set_allow_unassociated_targets(self, input: Option<bool>) -> Self

pub fn enabled(self, input: bool) -> Self

Whether the maintenance window is enabled.

pub fn set_enabled(self, input: Option<bool>) -> Self

pub fn replace(self, input: bool) -> Self

If True, then all fields that are required by the CreateMaintenanceWindow action are also required for this API request. Optional fields that are not specified are set to null.

pub fn set_replace(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<UpdateMaintenanceWindowInput, BuildError>

Consumes the builder and constructs a UpdateMaintenanceWindowInput