Module aws_sdk_ssm::model

Modules

account_sharing_info

See AccountSharingInfo

activation

See Activation

association

See Association

association_description

See AssociationDescription

association_execution

See AssociationExecution

association_execution_filter

See AssociationExecutionFilter

association_execution_target

See AssociationExecutionTarget

association_execution_targets_filter

See AssociationExecutionTargetsFilter

association_filter

See AssociationFilter

association_overview

See AssociationOverview

association_status

See AssociationStatus

association_version_info

See AssociationVersionInfo

attachment_content

See AttachmentContent

attachment_information

See AttachmentInformation

attachments_source

See AttachmentsSource

automation_execution

See AutomationExecution

automation_execution_filter

See AutomationExecutionFilter

automation_execution_metadata

See AutomationExecutionMetadata

baseline_override

See BaselineOverride

cloud_watch_output_config

See CloudWatchOutputConfig

command

See Command

command_filter

See CommandFilter

command_invocation

See CommandInvocation

command_plugin

See CommandPlugin

compliance_execution_summary

See ComplianceExecutionSummary

compliance_item

See ComplianceItem

compliance_item_entry

See ComplianceItemEntry

compliance_string_filter

See ComplianceStringFilter

compliance_summary_item

See ComplianceSummaryItem

compliant_summary

See CompliantSummary

create_association_batch_request_entry

See CreateAssociationBatchRequestEntry

describe_activations_filter

See DescribeActivationsFilter

document_default_version_description

See DocumentDefaultVersionDescription

document_description

See DocumentDescription

document_filter

See DocumentFilter

document_identifier

See DocumentIdentifier

document_key_values_filter

See DocumentKeyValuesFilter

document_metadata_response_info

See DocumentMetadataResponseInfo

document_parameter

See DocumentParameter

document_requires

See DocumentRequires

document_review_comment_source

See DocumentReviewCommentSource

document_reviewer_response_source

See DocumentReviewerResponseSource

document_reviews

See DocumentReviews

document_version_info

See DocumentVersionInfo

effective_patch

See EffectivePatch

failed_create_association

See FailedCreateAssociation

failure_details

See FailureDetails

instance_aggregated_association_overview

See InstanceAggregatedAssociationOverview

instance_association

See InstanceAssociation

instance_association_output_location

See InstanceAssociationOutputLocation

instance_association_output_url

See InstanceAssociationOutputUrl

instance_association_status_info

See InstanceAssociationStatusInfo

instance_information

See InstanceInformation

instance_information_filter

See InstanceInformationFilter

instance_information_string_filter

See InstanceInformationStringFilter

instance_patch_state

See InstancePatchState

instance_patch_state_filter

See InstancePatchStateFilter

inventory_aggregator

See InventoryAggregator

inventory_deletion_status_item

See InventoryDeletionStatusItem

inventory_deletion_summary

See InventoryDeletionSummary

inventory_deletion_summary_item

See InventoryDeletionSummaryItem

inventory_filter

See InventoryFilter

inventory_group

See InventoryGroup

inventory_item

See InventoryItem

inventory_item_attribute

See InventoryItemAttribute

inventory_item_schema

See InventoryItemSchema

inventory_result_entity

See InventoryResultEntity

inventory_result_item

See InventoryResultItem

logging_info

See LoggingInfo

maintenance_window_automation_parameters

See MaintenanceWindowAutomationParameters

maintenance_window_execution

See MaintenanceWindowExecution

maintenance_window_execution_task_identity

See MaintenanceWindowExecutionTaskIdentity

maintenance_window_execution_task_invocation_identity

See MaintenanceWindowExecutionTaskInvocationIdentity

maintenance_window_filter

See MaintenanceWindowFilter

maintenance_window_identity

See MaintenanceWindowIdentity

maintenance_window_identity_for_target

See MaintenanceWindowIdentityForTarget

maintenance_window_lambda_parameters

See MaintenanceWindowLambdaParameters

maintenance_window_run_command_parameters

See MaintenanceWindowRunCommandParameters

maintenance_window_step_functions_parameters

See MaintenanceWindowStepFunctionsParameters

maintenance_window_target

See MaintenanceWindowTarget

maintenance_window_task

See MaintenanceWindowTask

maintenance_window_task_invocation_parameters

See MaintenanceWindowTaskInvocationParameters

maintenance_window_task_parameter_value_expression

See MaintenanceWindowTaskParameterValueExpression

metadata_value

See MetadataValue

non_compliant_summary

See NonCompliantSummary

notification_config

See NotificationConfig

ops_aggregator

See OpsAggregator

ops_entity

See OpsEntity

ops_entity_item

See OpsEntityItem

ops_filter

See OpsFilter

ops_item

See OpsItem

ops_item_data_value

See OpsItemDataValue

ops_item_event_filter

See OpsItemEventFilter

ops_item_event_summary

See OpsItemEventSummary

ops_item_filter

See OpsItemFilter

ops_item_identity

See OpsItemIdentity

ops_item_notification

See OpsItemNotification

ops_item_related_item_summary

See OpsItemRelatedItemSummary

ops_item_related_items_filter

See OpsItemRelatedItemsFilter

ops_item_summary

See OpsItemSummary

ops_metadata

See OpsMetadata

ops_metadata_filter

See OpsMetadataFilter

ops_result_attribute

See OpsResultAttribute

output_source

See OutputSource

parameter

See Parameter

parameter_history

See ParameterHistory

parameter_inline_policy

See ParameterInlinePolicy

parameter_metadata

See ParameterMetadata

parameter_string_filter

See ParameterStringFilter

parameters_filter

See ParametersFilter

patch

See Patch

patch_baseline_identity

See PatchBaselineIdentity

patch_compliance_data

See PatchComplianceData

patch_filter

See PatchFilter

patch_filter_group

See PatchFilterGroup

patch_group_patch_baseline_mapping

See PatchGroupPatchBaselineMapping

patch_orchestrator_filter

See PatchOrchestratorFilter

patch_rule

See PatchRule

patch_rule_group

See PatchRuleGroup

patch_source

See PatchSource

patch_status

See PatchStatus

progress_counters

See ProgressCounters

related_ops_item

See RelatedOpsItem

resolved_targets

See ResolvedTargets

resource_compliance_summary_item

See ResourceComplianceSummaryItem

resource_data_sync_aws_organizations_source

See ResourceDataSyncAwsOrganizationsSource

resource_data_sync_destination_data_sharing

See ResourceDataSyncDestinationDataSharing

resource_data_sync_item

See ResourceDataSyncItem

resource_data_sync_organizational_unit

See ResourceDataSyncOrganizationalUnit

resource_data_sync_s3_destination

See ResourceDataSyncS3Destination

resource_data_sync_source

See ResourceDataSyncSource

resource_data_sync_source_with_state

See ResourceDataSyncSourceWithState

result_attribute

See ResultAttribute

review_information

See ReviewInformation

runbook

See Runbook

s3_output_location

See S3OutputLocation

s3_output_url

See S3OutputUrl

scheduled_window_execution

See ScheduledWindowExecution

service_setting

See ServiceSetting

session

See Session

session_filter

See SessionFilter

session_manager_output_url

See SessionManagerOutputUrl

severity_summary

See SeveritySummary

step_execution

See StepExecution

step_execution_filter

See StepExecutionFilter

tag

See Tag

target

See Target

target_location

See TargetLocation

Structs

AccountSharingInfo

Information includes the AWS account ID where the current document is shared and the version

Activation

An activation registers one or more on-premises servers or virtual machines (VMs) with AWS

Association

Describes an association of a Systems Manager document and an instance.

AssociationDescription

Describes the parameters for a document.

AssociationExecution

Includes information about the specified association.

AssociationExecutionFilter

Filters used in the request.

AssociationExecutionTarget

Includes information about the specified association execution.

AssociationExecutionTargetsFilter

Filters for the association execution.

AssociationFilter

Describes a filter.

AssociationOverview

Information about the association.

AssociationStatus

Describes an association status.

AssociationVersionInfo

Information about the association version.

AttachmentContent

A structure that includes attributes that describe a document attachment.

AttachmentInformation

An attribute of an attachment, such as the attachment name.

AttachmentsSource

Identifying information about a document attachment, including the file name and a key-value

AutomationExecution

Detailed information about the current state of an individual Automation execution.

AutomationExecutionFilter

A filter used to match specific automation executions. This is used to limit the scope of

AutomationExecutionMetadata

Details about a specific Automation execution.

BaselineOverride

Defines the basic information about a patch baseline override.

CloudWatchOutputConfig

Configuration options for sending command output to CloudWatch Logs.

Command

Describes a command request.

CommandFilter

Describes a command filter.

CommandInvocation

An invocation is copy of a command sent to a specific instance. A command can apply to one

CommandPlugin

Describes plugin details.

ComplianceExecutionSummary

A summary of the call execution that includes an execution ID, the type of execution (for

ComplianceItem

Information about the compliance as defined by the resource type. For example, for a patch

ComplianceItemEntry

Information about a compliance item.

ComplianceStringFilter

One or more filters. Use a filter to return a more specific list of results.

ComplianceSummaryItem

A summary of compliance information by compliance type.

CompliantSummary

A summary of resources that are compliant. The summary is organized according to the

CreateAssociationBatchRequestEntry

Describes the association of a Systems Manager SSM document and an instance.

DescribeActivationsFilter

Filter for the DescribeActivation API.

DocumentDefaultVersionDescription

A default version of a document.

DocumentDescription

Describes a Systems Manager document.

DocumentFilter

This data type is deprecated. Instead, use DocumentKeyValuesFilter.

DocumentIdentifier

Describes the name of a Systems Manager document.

DocumentKeyValuesFilter

One or more filters. Use a filter to return a more specific list of documents.

DocumentMetadataResponseInfo

Details about the response to a document review request.

DocumentParameter

Parameters specified in a System Manager document that run on the server when the command is

DocumentRequires

An SSM document required by the current document.

DocumentReviewCommentSource

Information about comments added to a document review request.

DocumentReviewerResponseSource

Information about a reviewer's response to a document review request.

DocumentReviews

Information about a document approval review.

DocumentVersionInfo

Version information about the document.

EffectivePatch

The EffectivePatch structure defines metadata about a patch along with the approval state of

FailedCreateAssociation

Describes a failed association.

FailureDetails

Information about an Automation failure.

InstanceAggregatedAssociationOverview

Status information about the aggregated associations.

InstanceAssociation

One or more association documents on the instance.

InstanceAssociationOutputLocation

An S3 bucket where you want to store the results of this request.

InstanceAssociationOutputUrl

The URL of S3 bucket where you want to store the results of this request.

InstanceAssociationStatusInfo

Status information about the instance association.

InstanceInformation

Describes a filter for a specific list of instances.

InstanceInformationFilter

Describes a filter for a specific list of instances. You can filter instances information by

InstanceInformationStringFilter

The filters to describe or get information about your managed instances.

InstancePatchState

Defines the high-level patch compliance state for a managed instance, providing information

InstancePatchStateFilter

Defines a filter used in DescribeInstancePatchStatesForPatchGroup used to

InventoryAggregator

Specifies the inventory type and attribute for the aggregation execution.

InventoryDeletionStatusItem

Status information returned by the DeleteInventory action.

InventoryDeletionSummary

Information about the delete operation.

InventoryDeletionSummaryItem

Either a count, remaining count, or a version number in a delete inventory summary.

InventoryFilter

One or more filters. Use a filter to return a more specific list of results.

InventoryGroup

A user-defined set of one or more filters on which to aggregate inventory data. Groups

InventoryItem

Information collected from managed instances based on your inventory policy document

InventoryItemAttribute

Attributes are the entries within the inventory item content. It contains name and

InventoryItemSchema

The inventory item schema definition. Users can use this to compose inventory query

InventoryResultEntity

Inventory query results.

InventoryResultItem

The inventory result item.

LoggingInfo

Information about an S3 bucket to write instance-level logs to.

MaintenanceWindowAutomationParameters

The parameters for an AUTOMATION task type.

MaintenanceWindowExecution

Describes the information about an execution of a maintenance window.

MaintenanceWindowExecutionTaskIdentity

Information about a task execution performed as part of a maintenance window

MaintenanceWindowExecutionTaskInvocationIdentity

Describes the information about a task invocation for a particular target as part of a task

MaintenanceWindowFilter

Filter used in the request. Supported filter keys are Name and Enabled.

MaintenanceWindowIdentity

Information about the maintenance window.

MaintenanceWindowIdentityForTarget

The maintenance window to which the specified target belongs.

MaintenanceWindowLambdaParameters

The parameters for a LAMBDA task type.

MaintenanceWindowRunCommandParameters

The parameters for a RUN_COMMAND task type.

MaintenanceWindowStepFunctionsParameters

The parameters for a STEP_FUNCTIONS task.

MaintenanceWindowTarget

The target registered with the maintenance window.

MaintenanceWindowTask

Information about a task defined for a maintenance window.

MaintenanceWindowTaskInvocationParameters

The parameters for task execution.

MaintenanceWindowTaskParameterValueExpression

Defines the values for a task parameter.

MetadataValue

Metadata to assign to an Application Manager application.

NonCompliantSummary

A summary of resources that are not compliant. The summary is organized according to

NotificationConfig

Configurations for sending notifications.

OpsAggregator

One or more aggregators for viewing counts of OpsItems using different dimensions such as

OpsEntity

The result of the query.

OpsEntityItem

The OpsItem summaries result item.

OpsFilter

A filter for viewing OpsItem summaries.

OpsItem

Operations engineers and IT professionals use OpsCenter to view, investigate, and remediate

OpsItemDataValue

An object that defines the value of the key and its type in the OperationalData map.

OpsItemEventFilter

Describes a filter for a specific list of OpsItem events. You can filter event information

OpsItemEventSummary

Summary information about an OpsItem event or that associated an OpsItem with a related

OpsItemFilter

Describes an OpsItem filter.

OpsItemIdentity

Information about the user or resource that created an OpsItem event.

OpsItemNotification

A notification about the OpsItem.

OpsItemRelatedItemSummary

Summary information about related-item resources for an OpsItem.

OpsItemRelatedItemsFilter

Describes a filter for a specific list of related-item resources.

OpsItemSummary

A count of OpsItems.

OpsMetadata

Operational metadata for an application in Application Manager.

OpsMetadataFilter

A filter to limit the number of OpsMetadata objects displayed.

OpsResultAttribute

The OpsItem data type to return.

OutputSource

Information about the source where the association execution details are stored.

Parameter

An Systems Manager parameter in Parameter Store.

ParameterHistory

Information about parameter usage.

ParameterInlinePolicy

One or more policies assigned to a parameter.

ParameterMetadata

Metadata includes information like the ARN of the last user and the date/time the parameter

ParameterStringFilter

One or more filters. Use a filter to return a more specific list of results.

ParametersFilter

This data type is deprecated. Instead, use ParameterStringFilter.

Patch

Represents metadata about a patch.

PatchBaselineIdentity

Defines the basic information about a patch baseline.

PatchComplianceData

Information about the state of a patch on a particular instance as it relates to the patch

PatchFilter

Defines which patches should be included in a patch baseline.

PatchFilterGroup

A set of patch filters, typically used for approval rules.

PatchGroupPatchBaselineMapping

The mapping between a patch group and the patch baseline the patch group is registered

PatchOrchestratorFilter

Defines a filter used in Patch Manager APIs.

PatchRule

Defines an approval rule for a patch baseline.

PatchRuleGroup

A set of rules defining the approval rules for a patch baseline.

PatchSource

Information about the patches to use to update the instances, including target operating

PatchStatus

Information about the approval status of a patch.

ProgressCounters

An aggregate of step execution statuses displayed in the AWS Console for a multi-Region and

RelatedOpsItem

An OpsItems that shares something in common with the current OpsItem. For example, related

ResolvedTargets

Information about targets that resolved during the Automation execution.

ResourceComplianceSummaryItem

Compliance summary information for a specific resource.

ResourceDataSyncAwsOrganizationsSource

Information about the AwsOrganizationsSource resource data sync source. A sync source of

ResourceDataSyncDestinationDataSharing

Synchronize Systems Manager Inventory data from multiple AWS accounts defined in AWS Organizations to

ResourceDataSyncItem

Information about a Resource Data Sync configuration, including its current status and last

ResourceDataSyncOrganizationalUnit

The AWS Organizations organizational unit data source for the sync.

ResourceDataSyncS3Destination

Information about the target S3 bucket for the Resource Data Sync.

ResourceDataSyncSource

Information about the source of the data included in the resource data sync.

ResourceDataSyncSourceWithState

The data type name for including resource data sync state. There are four sync

ResultAttribute

The inventory item result attribute.

ReviewInformation

Information about the result of a document review request.

Runbook

Information about an Automation runbook (Automation document) used in a runbook workflow in

S3OutputLocation

An S3 bucket where you want to store the results of this request.

S3OutputUrl

A URL for the S3 bucket where you want to store the results of this request.

ScheduledWindowExecution

Information about a scheduled execution for a maintenance window.

ServiceSetting

The service setting data structure.

Session

Information about a Session Manager connection to an instance.

SessionFilter

Describes a filter for Session Manager information.

SessionManagerOutputUrl

Reserved for future use.

SeveritySummary

The number of managed instances found for each patch severity level defined in the request

StepExecution

Detailed information about an the execution state of an Automation step.

StepExecutionFilter

A filter to limit the amount of step execution information returned by the call.

Tag

Metadata that you assign to your AWS resources. Tags enable you to categorize your resources

Target

An array of search criteria that targets instances using a Key,Value combination that you

TargetLocation

The combination of AWS Regions and accounts targeted by the current Automation

Enums

AssociationComplianceSeverity

AssociationExecutionFilterKey

AssociationExecutionTargetsFilterKey

AssociationFilterKey

AssociationFilterOperatorType

AssociationStatusName

AssociationSyncCompliance

AttachmentHashType

AttachmentsSourceKey

AutomationExecutionFilterKey

AutomationExecutionStatus

AutomationSubtype

AutomationType

CalendarState

CommandFilterKey

CommandInvocationStatus

CommandPluginStatus

CommandStatus

ComplianceQueryOperatorType

ComplianceSeverity

ComplianceStatus

ComplianceUploadType

ConnectionStatus

DescribeActivationsFilterKeys

DocumentFilterKey

DocumentFormat

DocumentHashType

DocumentMetadataEnum

DocumentParameterType

DocumentPermissionType

DocumentReviewAction

DocumentReviewCommentType

DocumentStatus

The status of a document.

DocumentType

ExecutionMode

Fault

NOTE: Fault::Unknown has been renamed to ::UnknownValue.

InstanceInformationFilterKey

InstancePatchStateOperatorType

InventoryAttributeDataType

InventoryDeletionStatus

InventoryQueryOperatorType

InventorySchemaDeleteOption

LastResourceDataSyncStatus

MaintenanceWindowExecutionStatus

MaintenanceWindowResourceType

MaintenanceWindowTaskType

NotificationEvent

NotificationType

OperatingSystem

OpsFilterOperatorType

OpsItemDataType

OpsItemEventFilterKey

OpsItemEventFilterOperator

OpsItemFilterKey

OpsItemFilterOperator

OpsItemRelatedItemsFilterKey

OpsItemRelatedItemsFilterOperator

OpsItemStatus

ParameterTier

ParameterType

ParametersFilterKey

PatchAction

PatchComplianceDataState

PatchComplianceLevel

PatchDeploymentStatus

PatchFilterKey

PatchOperationType

PatchProperty

PatchSet

PingStatus

PlatformType

RebootOption

ResourceDataSyncS3Format

ResourceType

ResourceTypeForTagging

ReviewStatus

SessionFilterKey

SessionState

SessionStatus

SignalType

StepExecutionFilterKey

StopType