Struct aws_sdk_ssm::model::association_description::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AssociationDescription

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the Systems Manager document.

pub fn set_name(self, input: Option<String>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the instance.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn association_version(self, input: impl Into<String>) -> Self

The association version.

pub fn set_association_version(self, input: Option<String>) -> Self

pub fn date(self, input: Instant) -> Self

The date when the association was made.

pub fn set_date(self, input: Option<Instant>) -> Self

pub fn last_update_association_date(self, input: Instant) -> Self

The date when the association was last updated.

pub fn set_last_update_association_date(self, input: Option<Instant>) -> Self

pub fn status(self, input: AssociationStatus) -> Self

The association status.

pub fn set_status(self, input: Option<AssociationStatus>) -> Self

pub fn overview(self, input: AssociationOverview) -> Self

Information about the association.

pub fn set_overview(self, input: Option<AssociationOverview>) -> Self

pub fn document_version(self, input: impl Into<String>) -> Self

The document version.

pub fn set_document_version(self, input: Option<String>) -> Self

pub fn automation_target_parameter_name(self, input: impl Into<String>) -> Self

Specify the target for the association. This target is required for associations that use an Automation document and target resources by using rate controls.

pub fn set_automation_target_parameter_name(self, input: Option<String>) -> Self

pub fn parameters(self, k: impl Into<String>, v: impl Into<Vec<String>>) -> Self

pub fn set_parameters(self, input: Option<HashMap<String, Vec<String>>>) -> Self

pub fn association_id(self, input: impl Into<String>) -> Self

The association ID.

pub fn set_association_id(self, input: Option<String>) -> Self

pub fn targets(self, input: impl Into<Target>) -> Self

pub fn set_targets(self, input: Option<Vec<Target>>) -> Self

pub fn schedule_expression(self, input: impl Into<String>) -> Self

A cron expression that specifies a schedule when the association runs.

pub fn set_schedule_expression(self, input: Option<String>) -> Self

pub fn output_location(self, input: InstanceAssociationOutputLocation) -> Self

An S3 bucket where you want to store the output details of the request.

pub fn set_output_location(
self,
input: Option<InstanceAssociationOutputLocation>
) -> Self

pub fn last_execution_date(self, input: Instant) -> Self

The date on which the association was last run.

pub fn set_last_execution_date(self, input: Option<Instant>) -> Self

pub fn last_successful_execution_date(self, input: Instant) -> Self

The last date on which the association was successfully run.

pub fn set_last_successful_execution_date(self, input: Option<Instant>) -> Self

pub fn association_name(self, input: impl Into<String>) -> Self

The association name.

pub fn set_association_name(self, input: Option<String>) -> Self

pub fn max_errors(self, input: impl Into<String>) -> Self

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

pub fn set_max_errors(self, input: Option<String>) -> Self

pub fn max_concurrency(self, input: impl Into<String>) -> Self

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

pub fn set_max_concurrency(self, input: Option<String>) -> Self

pub fn compliance_severity(self, input: AssociationComplianceSeverity) -> Self

The severity level that is assigned to the association.

pub fn set_compliance_severity(
self,
input: Option<AssociationComplianceSeverity>
) -> Self

pub fn sync_compliance(self, input: AssociationSyncCompliance) -> Self

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API action. In this case, compliance data is not managed by State Manager. It is managed by your direct call to the PutComplianceItems API action.

By default, all associations use AUTO mode.

pub fn set_sync_compliance(
self,
input: Option<AssociationSyncCompliance>
) -> Self

pub fn apply_only_at_cron_interval(self, input: bool) -> Self

By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter is not supported for rate expressions.

pub fn set_apply_only_at_cron_interval(self, input: Option<bool>) -> Self

pub fn calendar_names(self, input: impl Into<String>) -> Self

pub fn set_calendar_names(self, input: Option<Vec<String>>) -> Self

pub fn target_locations(self, input: impl Into<TargetLocation>) -> Self

pub fn set_target_locations(self, input: Option<Vec<TargetLocation>>) -> Self

pub fn build(self) -> AssociationDescription

Consumes the builder and constructs a AssociationDescription