Struct aws_sdk_ssm::model::automation_execution_metadata::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AutomationExecutionMetadata

Implementations

impl Builder

pub fn automation_execution_id(self, input: impl Into<String>) -> Self

The execution ID.

pub fn set_automation_execution_id(self, input: Option<String>) -> Self

pub fn document_name(self, input: impl Into<String>) -> Self

The name of the Automation document used during execution.

pub fn set_document_name(self, input: Option<String>) -> Self

pub fn document_version(self, input: impl Into<String>) -> Self

The document version used during the execution.

pub fn set_document_version(self, input: Option<String>) -> Self

pub fn automation_execution_status(
self,
input: AutomationExecutionStatus
) -> Self

The status of the execution.

pub fn set_automation_execution_status(
self,
input: Option<AutomationExecutionStatus>
) -> Self

pub fn execution_start_time(self, input: Instant) -> Self

The time the execution started.

pub fn set_execution_start_time(self, input: Option<Instant>) -> Self

pub fn execution_end_time(self, input: Instant) -> Self

The time the execution finished. This is not populated if the execution is still in progress.

pub fn set_execution_end_time(self, input: Option<Instant>) -> Self

pub fn executed_by(self, input: impl Into<String>) -> Self

The IAM role ARN of the user who ran the Automation.

pub fn set_executed_by(self, input: Option<String>) -> Self

pub fn log_file(self, input: impl Into<String>) -> Self

An S3 bucket where execution information is stored.

pub fn set_log_file(self, input: Option<String>) -> Self

pub fn outputs(self, k: impl Into<String>, v: impl Into<Vec<String>>) -> Self

pub fn set_outputs(self, input: Option<HashMap<String, Vec<String>>>) -> Self

pub fn mode(self, input: ExecutionMode) -> Self

The Automation execution mode.

pub fn set_mode(self, input: Option<ExecutionMode>) -> Self

pub fn parent_automation_execution_id(self, input: impl Into<String>) -> Self

The ExecutionId of the parent Automation.

pub fn set_parent_automation_execution_id(self, input: Option<String>) -> Self

pub fn current_step_name(self, input: impl Into<String>) -> Self

The name of the step that is currently running.

pub fn set_current_step_name(self, input: Option<String>) -> Self

pub fn current_action(self, input: impl Into<String>) -> Self

The action of the step that is currently running.

pub fn set_current_action(self, input: Option<String>) -> Self

pub fn failure_message(self, input: impl Into<String>) -> Self

The list of execution outputs as defined in the Automation document.

pub fn set_failure_message(self, input: Option<String>) -> Self

pub fn target_parameter_name(self, input: impl Into<String>) -> Self

The list of execution outputs as defined in the Automation document.

pub fn set_target_parameter_name(self, input: Option<String>) -> Self

pub fn targets(self, input: impl Into<Target>) -> Self

pub fn set_targets(self, input: Option<Vec<Target>>) -> Self

pub fn target_maps(self, input: impl Into<HashMap<String, Vec<String>>>) -> Self

pub fn set_target_maps(
self,
input: Option<Vec<HashMap<String, Vec<String>>>>
) -> Self

pub fn resolved_targets(self, input: ResolvedTargets) -> Self

A list of targets that resolved during the execution.

pub fn set_resolved_targets(self, input: Option<ResolvedTargets>) -> Self

pub fn max_concurrency(self, input: impl Into<String>) -> Self

The MaxConcurrency value specified by the user when starting the Automation.

pub fn set_max_concurrency(self, input: Option<String>) -> Self

pub fn max_errors(self, input: impl Into<String>) -> Self

The MaxErrors value specified by the user when starting the Automation.

pub fn set_max_errors(self, input: Option<String>) -> Self

pub fn target(self, input: impl Into<String>) -> Self

The list of execution outputs as defined in the Automation document.

pub fn set_target(self, input: Option<String>) -> Self

pub fn automation_type(self, input: AutomationType) -> Self

Use this filter with DescribeAutomationExecutions. Specify either Local or CrossAccount. CrossAccount is an Automation that runs in multiple AWS Regions and accounts. For more information, see Running Automation workflows in multiple AWS Regions and accounts in the AWS Systems Manager User Guide.

pub fn set_automation_type(self, input: Option<AutomationType>) -> Self

pub fn automation_subtype(self, input: AutomationSubtype) -> Self

The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

pub fn set_automation_subtype(self, input: Option<AutomationSubtype>) -> Self

pub fn scheduled_time(self, input: Instant) -> Self

The date and time the Automation operation is scheduled to start.

pub fn set_scheduled_time(self, input: Option<Instant>) -> Self

pub fn runbooks(self, input: impl Into<Runbook>) -> Self

pub fn set_runbooks(self, input: Option<Vec<Runbook>>) -> Self

pub fn ops_item_id(self, input: impl Into<String>) -> Self

The ID of an OpsItem that is created to represent a Change Manager change request.

pub fn set_ops_item_id(self, input: Option<String>) -> Self

pub fn association_id(self, input: impl Into<String>) -> Self

The ID of a State Manager association used in the Automation operation.

pub fn set_association_id(self, input: Option<String>) -> Self

pub fn change_request_name(self, input: impl Into<String>) -> Self

The name of the Change Manager change request.

pub fn set_change_request_name(self, input: Option<String>) -> Self

pub fn build(self) -> AutomationExecutionMetadata

Consumes the builder and constructs a AutomationExecutionMetadata