Struct aws_sdk_ssm::model::command_plugin::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CommandPlugin

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the plugin. Must be one of the following: aws:updateAgent, aws:domainjoin, aws:applications, aws:runPowerShellScript, aws:psmodule, aws:cloudWatch, aws:runShellScript, or aws:updateSSMAgent.

pub fn set_name(self, input: Option<String>) -> Self

pub fn status(self, input: CommandPluginStatus) -> Self

The status of this plugin. You can run a document with multiple plugins.

pub fn set_status(self, input: Option<CommandPluginStatus>) -> Self

pub fn status_details(self, input: impl Into<String>) -> Self

A detailed status of the plugin execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the AWS Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command has not been sent to the instance.

In Progress: The command has been sent to the instance but has not reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command was not delivered to the instance before the delivery timeout expired. Delivery timeouts do not count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution was not complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command was not successful on the instance. For a plugin, this indicates that the result code was not zero. For a command invocation, this indicates that the result code for one or more plugins was not zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist, or it might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit, and they don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

pub fn set_status_details(self, input: Option<String>) -> Self

pub fn response_code(self, input: i32) -> Self

A numeric response code generated after running the plugin.

pub fn set_response_code(self, input: Option<i32>) -> Self

pub fn response_start_date_time(self, input: Instant) -> Self

The time the plugin started running.

pub fn set_response_start_date_time(self, input: Option<Instant>) -> Self

pub fn response_finish_date_time(self, input: Instant) -> Self

The time the plugin stopped running. Could stop prematurely if, for example, a cancel command was sent.

pub fn set_response_finish_date_time(self, input: Option<Instant>) -> Self

pub fn output(self, input: impl Into<String>) -> Self

Output of the plugin execution.

pub fn set_output(self, input: Option<String>) -> Self

pub fn standard_output_url(self, input: impl Into<String>) -> Self

The URL for the complete text written by the plugin to stdout in Amazon S3. If the S3 bucket for the command was not specified, then this string is empty.

pub fn set_standard_output_url(self, input: Option<String>) -> Self

pub fn standard_error_url(self, input: impl Into<String>) -> Self

The URL for the complete text written by the plugin to stderr. If execution is not yet complete, then this string is empty.

pub fn set_standard_error_url(self, input: Option<String>) -> Self

pub fn output_s3_region(self, input: impl Into<String>) -> Self

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the S3 bucket region.

pub fn set_output_s3_region(self, input: Option<String>) -> Self

pub fn output_s3_bucket_name(self, input: impl Into<String>) -> Self

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

pub fn set_output_s3_bucket_name(self, input: Option<String>) -> Self

pub fn output_s3_key_prefix(self, input: impl Into<String>) -> Self

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

pub fn set_output_s3_key_prefix(self, input: Option<String>) -> Self

pub fn build(self) -> CommandPlugin

Consumes the builder and constructs a CommandPlugin