Struct aws_sdk_ssm::model::document_description::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DocumentDescription

Implementations

impl Builder

pub fn sha1(self, input: impl Into<String>) -> Self

The SHA1 hash of the document, which you can use for verification.

pub fn set_sha1(self, input: Option<String>) -> Self

pub fn hash(self, input: impl Into<String>) -> Self

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

pub fn set_hash(self, input: Option<String>) -> Self

pub fn hash_type(self, input: DocumentHashType) -> Self

The hash type of the document. Valid values include Sha256 or Sha1.

Sha1 hashes have been deprecated.

pub fn set_hash_type(self, input: Option<DocumentHashType>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name of the Systems Manager document.

pub fn set_name(self, input: Option<String>) -> Self

pub fn display_name(self, input: impl Into<String>) -> Self

The friendly name of the Systems Manager document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

pub fn set_display_name(self, input: Option<String>) -> Self

pub fn version_name(self, input: impl Into<String>) -> Self

The version of the artifact associated with the document.

pub fn set_version_name(self, input: Option<String>) -> Self

pub fn owner(self, input: impl Into<String>) -> Self

The AWS user account that created the document.

pub fn set_owner(self, input: Option<String>) -> Self

pub fn created_date(self, input: Instant) -> Self

The date when the document was created.

pub fn set_created_date(self, input: Option<Instant>) -> Self

pub fn status(self, input: DocumentStatus) -> Self

The status of the Systems Manager document.

pub fn set_status(self, input: Option<DocumentStatus>) -> Self

pub fn status_information(self, input: impl Into<String>) -> Self

A message returned by AWS Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket does not exist. Verify that the URL of the S3 bucket is correct."

pub fn set_status_information(self, input: Option<String>) -> Self

pub fn document_version(self, input: impl Into<String>) -> Self

The document version.

pub fn set_document_version(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description of the document.

pub fn set_description(self, input: Option<String>) -> Self

pub fn parameters(self, input: impl Into<DocumentParameter>) -> Self

pub fn set_parameters(self, input: Option<Vec<DocumentParameter>>) -> Self

pub fn platform_types(self, input: impl Into<PlatformType>) -> Self

pub fn set_platform_types(self, input: Option<Vec<PlatformType>>) -> Self

pub fn document_type(self, input: DocumentType) -> Self

The type of document.

pub fn set_document_type(self, input: Option<DocumentType>) -> Self

pub fn schema_version(self, input: impl Into<String>) -> Self

The schema version.

pub fn set_schema_version(self, input: Option<String>) -> Self

pub fn latest_version(self, input: impl Into<String>) -> Self

The latest version of the document.

pub fn set_latest_version(self, input: Option<String>) -> Self

pub fn default_version(self, input: impl Into<String>) -> Self

The default version.

pub fn set_default_version(self, input: Option<String>) -> Self

pub fn document_format(self, input: DocumentFormat) -> Self

The document format, either JSON or YAML.

pub fn set_document_format(self, input: Option<DocumentFormat>) -> Self

pub fn target_type(self, input: impl Into<String>) -> Self

The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see AWS resource and property types reference in the AWS CloudFormation User Guide.

pub fn set_target_type(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn attachments_information(
self,
input: impl Into<AttachmentInformation>
) -> Self

pub fn set_attachments_information(
self,
input: Option<Vec<AttachmentInformation>>
) -> Self

pub fn requires(self, input: impl Into<DocumentRequires>) -> Self

pub fn set_requires(self, input: Option<Vec<DocumentRequires>>) -> Self

pub fn author(self, input: impl Into<String>) -> Self

The user in your organization who created the document.

pub fn set_author(self, input: Option<String>) -> Self

pub fn review_information(self, input: impl Into<ReviewInformation>) -> Self

pub fn set_review_information(
self,
input: Option<Vec<ReviewInformation>>
) -> Self

pub fn approved_version(self, input: impl Into<String>) -> Self

The version of the document currently approved for use in the organization.

pub fn set_approved_version(self, input: Option<String>) -> Self

pub fn pending_review_version(self, input: impl Into<String>) -> Self

The version of the document that is currently under review.

pub fn set_pending_review_version(self, input: Option<String>) -> Self

pub fn review_status(self, input: ReviewStatus) -> Self

The current status of the review.

pub fn set_review_status(self, input: Option<ReviewStatus>) -> Self

pub fn build(self) -> DocumentDescription

Consumes the builder and constructs a DocumentDescription