Struct aws_sdk_ssm::model::instance_information::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for InstanceInformation

Implementations

impl Builder

pub fn instance_id(self, input: impl Into<String>) -> Self

The instance ID.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn ping_status(self, input: PingStatus) -> Self

Connection status of SSM Agent.

The status Inactive has been deprecated and is no longer in use.

pub fn set_ping_status(self, input: Option<PingStatus>) -> Self

pub fn last_ping_date_time(self, input: Instant) -> Self

The date and time when the agent last pinged the Systems Manager service.

pub fn set_last_ping_date_time(self, input: Option<Instant>) -> Self

pub fn agent_version(self, input: impl Into<String>) -> Self

The version of SSM Agent running on your Linux instance.

pub fn set_agent_version(self, input: Option<String>) -> Self

pub fn is_latest_version(self, input: bool) -> Self

Indicates whether the latest version of SSM Agent is running on your Linux Managed Instance. This field does not indicate whether or not the latest version is installed on Windows managed instances, because some older versions of Windows Server use the EC2Config service to process SSM requests.

pub fn set_is_latest_version(self, input: Option<bool>) -> Self

pub fn platform_type(self, input: PlatformType) -> Self

The operating system platform type.

pub fn set_platform_type(self, input: Option<PlatformType>) -> Self

pub fn platform_name(self, input: impl Into<String>) -> Self

The name of the operating system platform running on your instance.

pub fn set_platform_name(self, input: Option<String>) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

The version of the OS platform running on your instance.

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn activation_id(self, input: impl Into<String>) -> Self

The activation ID created by Systems Manager when the server or VM was registered.

pub fn set_activation_id(self, input: Option<String>) -> Self

pub fn iam_role(self, input: impl Into<String>) -> Self

The Amazon Identity and Access Management (IAM) role assigned to the on-premises Systems Manager managed instance. This call does not return the IAM role for EC2 instances. To retrieve the IAM role for an EC2 instance, use the Amazon EC2 DescribeInstances action. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the AWS CLI Command Reference.

pub fn set_iam_role(self, input: Option<String>) -> Self

pub fn registration_date(self, input: Instant) -> Self

The date the server or VM was registered with AWS as a managed instance.

pub fn set_registration_date(self, input: Option<Instant>) -> Self

pub fn resource_type(self, input: ResourceType) -> Self

The type of instance. Instances are either EC2 instances or managed instances.

pub fn set_resource_type(self, input: Option<ResourceType>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name assigned to an on-premises server or virtual machine (VM) when it is activated as a Systems Manager managed instance. The name is specified as the DefaultInstanceName property using the CreateActivation command. It is applied to the managed instance by specifying the Activation Code and Activation ID when you install SSM Agent on the instance, as explained in Install SSM Agent for a hybrid environment (Linux) and Install SSM Agent for a hybrid environment (Windows). To retrieve the Name tag of an EC2 instance, use the Amazon EC2 DescribeInstances action. For information, see DescribeInstances in the Amazon EC2 API Reference or describe-instances in the AWS CLI Command Reference.

pub fn set_name(self, input: Option<String>) -> Self

pub fn ip_address(self, input: impl Into<String>) -> Self

The IP address of the managed instance.

pub fn set_ip_address(self, input: Option<String>) -> Self

pub fn computer_name(self, input: impl Into<String>) -> Self

The fully qualified host name of the managed instance.

pub fn set_computer_name(self, input: Option<String>) -> Self

pub fn association_status(self, input: impl Into<String>) -> Self

The status of the association.

pub fn set_association_status(self, input: Option<String>) -> Self

pub fn last_association_execution_date(self, input: Instant) -> Self

The date the association was last run.

pub fn set_last_association_execution_date(self, input: Option<Instant>) -> Self

pub fn last_successful_association_execution_date(self, input: Instant) -> Self

The last date the association was successfully run.

pub fn set_last_successful_association_execution_date(
self,
input: Option<Instant>
) -> Self

pub fn association_overview(
self,
input: InstanceAggregatedAssociationOverview
) -> Self

Information about the association.

pub fn set_association_overview(
self,
input: Option<InstanceAggregatedAssociationOverview>
) -> Self

pub fn build(self) -> InstanceInformation

Consumes the builder and constructs a InstanceInformation