Struct aws_sdk_ssm::model::maintenance_window_identity::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for MaintenanceWindowIdentity

Implementations

impl Builder

pub fn window_id(self, inp: impl Into<String>) -> Self

The ID of the maintenance window.

pub fn set_window_id(self, inp: Option<String>) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

The name of the maintenance window.

pub fn set_name(self, inp: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the maintenance window.

pub fn set_description(self, inp: Option<String>) -> Self

pub fn enabled(self, inp: bool) -> Self

Indicates whether the maintenance window is enabled.

pub fn set_enabled(self, inp: bool) -> Self

pub fn duration(self, inp: i32) -> Self

The duration of the maintenance window in hours.

pub fn set_duration(self, inp: i32) -> Self

pub fn cutoff(self, inp: i32) -> Self

The number of hours before the end of the maintenance window that Systems Manager stops scheduling new tasks for execution.

pub fn set_cutoff(self, inp: i32) -> Self

pub fn schedule(self, inp: impl Into<String>) -> Self

The schedule of the maintenance window in the form of a cron or rate expression.

pub fn set_schedule(self, inp: Option<String>) -> Self

pub fn schedule_timezone(self, inp: impl Into<String>) -> Self

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format.

pub fn set_schedule_timezone(self, inp: Option<String>) -> Self

pub fn schedule_offset(self, inp: i32) -> Self

The number of days to wait to run a maintenance window after the scheduled CRON expression date and time.

pub fn set_schedule_offset(self, inp: Option<i32>) -> Self

pub fn end_date(self, inp: impl Into<String>) -> Self

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive.

pub fn set_end_date(self, inp: Option<String>) -> Self

pub fn start_date(self, inp: impl Into<String>) -> Self

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active.

pub fn set_start_date(self, inp: Option<String>) -> Self

pub fn next_execution_time(self, inp: impl Into<String>) -> Self

The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

pub fn set_next_execution_time(self, inp: Option<String>) -> Self

pub fn build(self) -> MaintenanceWindowIdentity

Consumes the builder and constructs a MaintenanceWindowIdentity