Struct aws_sdk_ssm::model::AutomationExecution

#[non_exhaustive]
pub struct AutomationExecution {

    pub automation_execution_id: Option<String>,
    pub document_name: Option<String>,
    pub document_version: Option<String>,
    pub execution_start_time: Option<Instant>,
    pub execution_end_time: Option<Instant>,
    pub automation_execution_status: Option<AutomationExecutionStatus>,
    pub step_executions: Option<Vec<StepExecution>>,
    pub step_executions_truncated: bool,
    pub parameters: Option<HashMap<String, Vec<String>>>,
    pub outputs: Option<HashMap<String, Vec<String>>>,
    pub failure_message: Option<String>,
    pub mode: Option<ExecutionMode>,
    pub parent_automation_execution_id: Option<String>,
    pub executed_by: Option<String>,
    pub current_step_name: Option<String>,
    pub current_action: Option<String>,
    pub target_parameter_name: Option<String>,
    pub targets: Option<Vec<Target>>,
    pub target_maps: Option<Vec<HashMap<String, Vec<String>>>>,
    pub resolved_targets: Option<ResolvedTargets>,
    pub max_concurrency: Option<String>,
    pub max_errors: Option<String>,
    pub target: Option<String>,
    pub target_locations: Option<Vec<TargetLocation>>,
    pub progress_counters: Option<ProgressCounters>,
    pub automation_subtype: Option<AutomationSubtype>,
    pub scheduled_time: Option<Instant>,
    pub runbooks: Option<Vec<Runbook>>,
    pub ops_item_id: Option<String>,
    pub association_id: Option<String>,
    pub change_request_name: Option<String>,

}

Detailed information about the current state of an individual Automation execution.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

automation_execution_id: Option<String>

The execution ID.

document_name: Option<String>

The name of the Automation document used during the execution.

document_version: Option<String>

The version of the document to use during execution.

execution_start_time: Option<Instant>

The time the execution started.

execution_end_time: Option<Instant>

The time the execution finished.

automation_execution_status: Option<AutomationExecutionStatus>

The execution status of the Automation.

step_executions: Option<Vec<StepExecution>>

A list of details about the current state of all steps that comprise an execution. An Automation document contains a list of steps that are run in order.

step_executions_truncated: bool

A boolean value that indicates if the response contains the full list of the Automation step executions. If true, use the DescribeAutomationStepExecutions API action to get the full list of step executions.

parameters: Option<HashMap<String, Vec<String>>>

The key-value map of execution parameters, which were supplied when calling StartAutomationExecution.

outputs: Option<HashMap<String, Vec<String>>>

The list of execution outputs as defined in the automation document.

failure_message: Option<String>

A message describing why an execution has failed, if the status is set to Failed.

mode: Option<ExecutionMode>

The automation execution mode.

parent_automation_execution_id: Option<String>

The AutomationExecutionId of the parent automation.

executed_by: Option<String>

The Amazon Resource Name (ARN) of the user who ran the automation.

current_step_name: Option<String>

The name of the step that is currently running.

current_action: Option<String>

The action of the step that is currently running.

target_parameter_name: Option<String>

The parameter name.

targets: Option<Vec<Target>>

The specified targets.

target_maps: Option<Vec<HashMap<String, Vec<String>>>>

The specified key-value mapping of document parameters to target resources.

resolved_targets: Option<ResolvedTargets>

A list of resolved targets in the rate control execution.

max_concurrency: Option<String>

The MaxConcurrency value specified by the user when the execution started.

max_errors: Option<String>

The MaxErrors value specified by the user when the execution started.

target: Option<String>

The target of the execution.

target_locations: Option<Vec<TargetLocation>>

The combination of AWS Regions and/or AWS accounts where you want to run the Automation.

progress_counters: Option<ProgressCounters>

An aggregate of step execution statuses displayed in the AWS Console for a multi-Region and multi-account Automation execution.

automation_subtype: Option<AutomationSubtype>

The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

scheduled_time: Option<Instant>

The date and time the Automation operation is scheduled to start.

runbooks: Option<Vec<Runbook>>

Information about the Automation runbooks (Automation documents) that are run as part of a runbook workflow.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

ops_item_id: Option<String>

The ID of an OpsItem that is created to represent a Change Manager change request.

association_id: Option<String>

The ID of a State Manager association used in the Automation operation.

change_request_name: Option<String>

The name of the Change Manager change request.

Implementations

impl AutomationExecution

pub fn builder() -> Builder

Creates a new builder-style object to manufacture AutomationExecution