Struct aws_sdk_ssm::model::CommandInvocation

#[non_exhaustive]
pub struct CommandInvocation {

    pub command_id: Option<String>,
    pub instance_id: Option<String>,
    pub instance_name: Option<String>,
    pub comment: Option<String>,
    pub document_name: Option<String>,
    pub document_version: Option<String>,
    pub requested_date_time: Option<Instant>,
    pub status: Option<CommandInvocationStatus>,
    pub status_details: Option<String>,
    pub trace_output: Option<String>,
    pub standard_output_url: Option<String>,
    pub standard_error_url: Option<String>,
    pub command_plugins: Option<Vec<CommandPlugin>>,
    pub service_role: Option<String>,
    pub notification_config: Option<NotificationConfig>,
    pub cloud_watch_output_config: Option<CloudWatchOutputConfig>,

}

An invocation is copy of a command sent to a specific instance. A command can apply to one or more instances. A command invocation applies to one instance. For example, if a user runs SendCommand against three instances, then a command invocation is created for each requested instance ID. A command invocation returns status and detail information about a command you ran.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

command_id: Option<String>

The command against which this invocation was requested.

instance_id: Option<String>

The instance ID in which this invocation was requested.

instance_name: Option<String>

The name of the invocation target. For EC2 instances this is the value for the aws:Name tag. For on-premises instances, this is the name of the instance.

comment: Option<String>

User-specified information about the command, such as a brief description of what the command should do.

document_name: Option<String>

The document name that was requested for execution.

document_version: Option<String>

The SSM document version.

requested_date_time: Option<Instant>

The time and date the request was sent to this instance.

status: Option<CommandInvocationStatus>

Whether or not the invocation succeeded, failed, or is pending.

status_details: Option<String>

A detailed status of the command execution for each invocation (each instance targeted by the command). StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the AWS Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command has not been sent to the instance.

In Progress: The command has been sent to the instance but has not reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command was not delivered to the instance before the delivery timeout expired. Delivery timeouts do not count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution was not complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command was not successful on the instance. For a plugin, this indicates that the result code was not zero. For a command invocation, this indicates that the result code for one or more plugins was not zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

trace_output: Option<String>

Gets the trace output sent by the agent.

standard_output_url: Option<String>

The URL to the plugin's StdOut file in Amazon S3, if the S3 bucket was defined for the parent command. For an invocation, StandardOutputUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

standard_error_url: Option<String>

The URL to the plugin's StdErr file in Amazon S3, if the S3 bucket was defined for the parent command. For an invocation, StandardErrorUrl is populated if there is just one plugin defined for the command, and the S3 bucket was defined for the command.

command_plugins: Option<Vec<CommandPlugin>>

Plugins processed by the command.

service_role: Option<String>

The IAM service role that Run Command uses to act on your behalf when sending notifications about command status changes on a per instance basis.

notification_config: Option<NotificationConfig>

Configurations for sending notifications about command status changes on a per instance basis.

cloud_watch_output_config: Option<CloudWatchOutputConfig>

CloudWatch Logs information where you want Systems Manager to send the command output.

Implementations

impl CommandInvocation

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CommandInvocation