Struct aws_sdk_ssm::model::CommandPlugin

#[non_exhaustive]
pub struct CommandPlugin {
    pub name: Option<String>,
    pub status: Option<CommandPluginStatus>,
    pub status_details: Option<String>,
    pub response_code: i32,
    pub response_start_date_time: Option<Instant>,
    pub response_finish_date_time: Option<Instant>,
    pub output: Option<String>,
    pub standard_output_url: Option<String>,
    pub standard_error_url: Option<String>,
    pub output_s3_region: Option<String>,
    pub output_s3_bucket_name: Option<String>,
    pub output_s3_key_prefix: Option<String>,
}

Describes plugin details.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The name of the plugin. Must be one of the following: aws:updateAgent, aws:domainjoin, aws:applications, aws:runPowerShellScript, aws:psmodule, aws:cloudWatch, aws:runShellScript, or aws:updateSSMAgent.

status: Option<CommandPluginStatus>

The status of this plugin. You can run a document with multiple plugins.

status_details: Option<String>

A detailed status of the plugin execution. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Success: The execution of the command or plugin was successfully completed. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: Command execution started on the instance, but the execution wasn't complete before the execution timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't successful on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist, or it might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit, and they don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

response_code: i32

A numeric response code generated after running the plugin.

response_start_date_time: Option<Instant>

The time the plugin started running.

response_finish_date_time: Option<Instant>

The time the plugin stopped running. Could stop prematurely if, for example, a cancel command was sent.

output: Option<String>

Output of the plugin execution.

standard_output_url: Option<String>

The URL for the complete text written by the plugin to stdout in Amazon S3. If the S3 bucket for the command wasn't specified, then this string is empty.

standard_error_url: Option<String>

The URL for the complete text written by the plugin to stderr. If execution isn't yet complete, then this string is empty.

output_s3_region: Option<String>

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Amazon Web Services Systems Manager automatically determines the S3 bucket region.

output_s3_bucket_name: Option<String>

The S3 bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

output_s3_key_prefix: Option<String>

The S3 directory path inside the bucket where the responses to the command executions should be stored. This was requested when issuing the command. For example, in the following response:

doc-example-bucket/ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix/i-02573cafcfEXAMPLE/awsrunShellScript

doc-example-bucket is the name of the S3 bucket;

ab19cb99-a030-46dd-9dfc-8eSAMPLEPre-Fix is the name of the S3 prefix;

i-02573cafcfEXAMPLE is the instance ID;

awsrunShellScript is the name of the plugin.

Implementations

impl CommandPlugin

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CommandPlugin