Struct aws_sdk_ssm::model::DocumentDescription

#[non_exhaustive]
pub struct DocumentDescription {

    pub sha1: Option<String>,
    pub hash: Option<String>,
    pub hash_type: Option<DocumentHashType>,
    pub name: Option<String>,
    pub display_name: Option<String>,
    pub version_name: Option<String>,
    pub owner: Option<String>,
    pub created_date: Option<Instant>,
    pub status: Option<DocumentStatus>,
    pub status_information: Option<String>,
    pub document_version: Option<String>,
    pub description: Option<String>,
    pub parameters: Option<Vec<DocumentParameter>>,
    pub platform_types: Option<Vec<PlatformType>>,
    pub document_type: Option<DocumentType>,
    pub schema_version: Option<String>,
    pub latest_version: Option<String>,
    pub default_version: Option<String>,
    pub document_format: Option<DocumentFormat>,
    pub target_type: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub attachments_information: Option<Vec<AttachmentInformation>>,
    pub requires: Option<Vec<DocumentRequires>>,
    pub author: Option<String>,
    pub review_information: Option<Vec<ReviewInformation>>,
    pub approved_version: Option<String>,
    pub pending_review_version: Option<String>,
    pub review_status: Option<ReviewStatus>,

}

Describes a Systems Manager document.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

sha1: Option<String>

The SHA1 hash of the document, which you can use for verification.

hash: Option<String>

The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.

hash_type: Option<DocumentHashType>

The hash type of the document. Valid values include Sha256 or Sha1.

Sha1 hashes have been deprecated.

name: Option<String>

The name of the Systems Manager document.

display_name: Option<String>

The friendly name of the Systems Manager document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

version_name: Option<String>

The version of the artifact associated with the document.

owner: Option<String>

The AWS user account that created the document.

created_date: Option<Instant>

The date when the document was created.

status: Option<DocumentStatus>

The status of the Systems Manager document.

status_information: Option<String>

A message returned by AWS Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket does not exist. Verify that the URL of the S3 bucket is correct."

document_version: Option<String>

The document version.

description: Option<String>

A description of the document.

parameters: Option<Vec<DocumentParameter>>

A description of the parameters for a document.

platform_types: Option<Vec<PlatformType>>

The list of OS platforms compatible with this Systems Manager document.

document_type: Option<DocumentType>

The type of document.

schema_version: Option<String>

The schema version.

latest_version: Option<String>

The latest version of the document.

default_version: Option<String>

The default version.

document_format: Option<DocumentFormat>

The document format, either JSON or YAML.

target_type: Option<String>

The target type which defines the kinds of resources the document can run on. For example, /AWS::EC2::Instance. For a list of valid resource types, see AWS resource and property types reference in the AWS CloudFormation User Guide.

tags: Option<Vec<Tag>>

The tags, or metadata, that have been applied to the document.

attachments_information: Option<Vec<AttachmentInformation>>

Details about the document attachments, including names, locations, sizes, and so on.

requires: Option<Vec<DocumentRequires>>

A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

author: Option<String>

The user in your organization who created the document.

review_information: Option<Vec<ReviewInformation>>

Details about the review of a document.

approved_version: Option<String>

The version of the document currently approved for use in the organization.

pending_review_version: Option<String>

The version of the document that is currently under review.

review_status: Option<ReviewStatus>

The current status of the review.

Implementations

impl DocumentDescription

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DocumentDescription