Struct aws_sdk_ssm::model::InstancePatchState

#[non_exhaustive]
pub struct InstancePatchState {

    pub instance_id: Option<String>,
    pub patch_group: Option<String>,
    pub baseline_id: Option<String>,
    pub snapshot_id: Option<String>,
    pub install_override_list: Option<String>,
    pub owner_information: Option<String>,
    pub installed_count: i32,
    pub installed_other_count: i32,
    pub installed_pending_reboot_count: Option<i32>,
    pub installed_rejected_count: Option<i32>,
    pub missing_count: i32,
    pub failed_count: i32,
    pub unreported_not_applicable_count: Option<i32>,
    pub not_applicable_count: i32,
    pub operation_start_time: Option<Instant>,
    pub operation_end_time: Option<Instant>,
    pub operation: Option<PatchOperationType>,
    pub last_no_reboot_install_operation_time: Option<Instant>,
    pub reboot_option: Option<RebootOption>,
    pub critical_non_compliant_count: Option<i32>,
    pub security_non_compliant_count: Option<i32>,
    pub other_non_compliant_count: Option<i32>,

}

Defines the high-level patch compliance state for a managed instance, providing information about the number of installed, missing, not applicable, and failed patches along with metadata about the operation when this information was gathered for the instance.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

instance_id: Option<String>

The ID of the managed instance the high-level patch compliance information was collected for.

patch_group: Option<String>

The name of the patch group the managed instance belongs to.

baseline_id: Option<String>

The ID of the patch baseline used to patch the instance.

snapshot_id: Option<String>

The ID of the patch baseline snapshot used during the patching operation when this compliance data was collected.

install_override_list: Option<String>

An https URL or an Amazon S3 path-style URL to a list of patches to be installed. This patch installation list, which you maintain in an S3 bucket in YAML format and specify in the SSM document AWS-RunPatchBaseline, overrides the patches specified by the default patch baseline.

For more information about the InstallOverrideList parameter, see About the SSM document AWS-RunPatchBaseline in the AWS Systems Manager User Guide.

owner_information: Option<String>

Placeholder information. This field will always be empty in the current release of the service.

installed_count: i32

The number of patches from the patch baseline that are installed on the instance.

installed_other_count: i32

The number of patches not specified in the patch baseline that are installed on the instance.

installed_pending_reboot_count: Option<i32>

The number of patches installed by Patch Manager since the last time the instance was rebooted.

installed_rejected_count: Option<i32>

The number of patches installed on an instance that are specified in a RejectedPatches list. Patches with a status of InstalledRejected were typically installed before they were added to a RejectedPatches list.

If ALLOW_AS_DEPENDENCY is the specified option for RejectedPatchesAction, the value of InstalledRejectedCount will always be 0 (zero).

missing_count: i32

The number of patches from the patch baseline that are applicable for the instance but aren't currently installed.

failed_count: i32

The number of patches from the patch baseline that were attempted to be installed during the last patching operation, but failed to install.

unreported_not_applicable_count: Option<i32>

The number of patches beyond the supported limit of NotApplicableCount that are not reported by name to Systems Manager Inventory.

not_applicable_count: i32

The number of patches from the patch baseline that aren't applicable for the instance and therefore aren't installed on the instance. This number may be truncated if the list of patch names is very large. The number of patches beyond this limit are reported in UnreportedNotApplicableCount.

operation_start_time: Option<Instant>

The time the most recent patching operation was started on the instance.

operation_end_time: Option<Instant>

The time the most recent patching operation completed on the instance.

operation: Option<PatchOperationType>

The type of patching operation that was performed: SCAN (assess patch compliance state) or INSTALL (install missing patches).

last_no_reboot_install_operation_time: Option<Instant>

The time of the last attempt to patch the instance with NoReboot specified as the reboot option.

reboot_option: Option<RebootOption>

Indicates the reboot option specified in the patch baseline.

Reboot options apply to Install operations only. Reboots are not attempted for Patch Manager Scan operations.

RebootIfNeeded: Patch Manager tries to reboot the instance if it installed any patches, or if any patches are detected with a status of InstalledPendingReboot.

NoReboot: Patch Manager attempts to install missing packages without trying to reboot the system. Patches installed with this option are assigned a status of InstalledPendingReboot. These patches might not be in effect until a reboot is performed.

critical_non_compliant_count: Option<i32>

The number of instances where patches that are specified as "Critical" for compliance reporting in the patch baseline are not installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

security_non_compliant_count: Option<i32>

The number of instances where patches that are specified as "Security" in a patch advisory are not installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

other_non_compliant_count: Option<i32>

The number of instances with patches installed that are specified as other than "Critical" or "Security" but are not compliant with the patch baseline. The status of these instances is NON_COMPLIANT.

Implementations

impl InstancePatchState

pub fn builder() -> Builder

Creates a new builder-style object to manufacture InstancePatchState