Struct aws_sdk_ssm::model::OpsItem

#[non_exhaustive]
pub struct OpsItem {

    pub created_by: Option<String>,
    pub ops_item_type: Option<String>,
    pub created_time: Option<Instant>,
    pub description: Option<String>,
    pub last_modified_by: Option<String>,
    pub last_modified_time: Option<Instant>,
    pub notifications: Option<Vec<OpsItemNotification>>,
    pub priority: Option<i32>,
    pub related_ops_items: Option<Vec<RelatedOpsItem>>,
    pub status: Option<OpsItemStatus>,
    pub ops_item_id: Option<String>,
    pub version: Option<String>,
    pub title: Option<String>,
    pub source: Option<String>,
    pub operational_data: Option<HashMap<String, OpsItemDataValue>>,
    pub category: Option<String>,
    pub severity: Option<String>,
    pub actual_start_time: Option<Instant>,
    pub actual_end_time: Option<Instant>,
    pub planned_start_time: Option<Instant>,
    pub planned_end_time: Option<Instant>,

}

Operations engineers and IT professionals use Amazon Web Services Systems Manager OpsCenter to view, investigate, and remediate operational work items (OpsItems) impacting the performance and health of their Amazon Web Services resources. OpsCenter is integrated with Amazon EventBridge and Amazon CloudWatch. This means you can configure these services to automatically create an OpsItem in OpsCenter when a CloudWatch alarm enters the ALARM state or when EventBridge processes an event from any Amazon Web Services service that publishes events. Configuring Amazon CloudWatch alarms and EventBridge events to automatically create OpsItems allows you to quickly diagnose and remediate issues with Amazon Web Services resources from a single console.

To help you diagnose issues, each OpsItem includes contextually relevant information such as the name and ID of the Amazon Web Services resource that generated the OpsItem, alarm or event details, alarm history, and an alarm timeline graph. For the Amazon Web Services resource, OpsCenter aggregates information from Config, CloudTrail logs, and EventBridge, so you don't have to navigate across multiple console pages during your investigation. For more information, see OpsCenter in the Amazon Web Services Systems Manager User Guide.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

created_by: Option<String>

The ARN of the Amazon Web Services account that created the OpsItem.

ops_item_type: Option<String>

The type of OpsItem. Currently, the only valid values are /aws/changerequest and /aws/issue.

created_time: Option<Instant>

The date and time the OpsItem was created.

description: Option<String>

The OpsItem description.

last_modified_by: Option<String>

The ARN of the Amazon Web Services account that last updated the OpsItem.

last_modified_time: Option<Instant>

The date and time the OpsItem was last updated.

notifications: Option<Vec<OpsItemNotification>>

The Amazon Resource Name (ARN) of an Amazon Simple Notification Service (Amazon SNS) topic where notifications are sent when this OpsItem is edited or changed.

priority: Option<i32>

The importance of this OpsItem in relation to other OpsItems in the system.

related_ops_items: Option<Vec<RelatedOpsItem>>

One or more OpsItems that share something in common with the current OpsItem. For example, related OpsItems can include OpsItems with similar error messages, impacted resources, or statuses for the impacted resource.

status: Option<OpsItemStatus>

The OpsItem status. Status can be Open, In Progress, or Resolved. For more information, see Editing OpsItem details in the Amazon Web Services Systems Manager User Guide.

ops_item_id: Option<String>

The ID of the OpsItem.

version: Option<String>

The version of this OpsItem. Each time the OpsItem is edited the version number increments by one.

title: Option<String>

A short heading that describes the nature of the OpsItem and the impacted resource.

source: Option<String>

The origin of the OpsItem, such as Amazon EC2 or Systems Manager. The impacted resource is a subset of source.

operational_data: Option<HashMap<String, OpsItemDataValue>>

Operational data is custom data that provides useful reference details about the OpsItem. For example, you can specify log files, error strings, license keys, troubleshooting tips, or other relevant data. You enter operational data as key-value pairs. The key has a maximum length of 128 characters. The value has a maximum size of 20 KB.

Operational data keys can't begin with the following: amazon, aws, amzn, ssm, /amazon, /aws, /amzn, /ssm.

You can choose to make the data searchable by other users in the account or you can restrict search access. Searchable data means that all users with access to the OpsItem Overview page (as provided by the DescribeOpsItems API operation) can view and search on the specified data. Operational data that isn't searchable is only viewable by users who have access to the OpsItem (as provided by the GetOpsItem API operation).

Use the /aws/resources key in OperationalData to specify a related resource in the request. Use the /aws/automations key in OperationalData to associate an Automation runbook with the OpsItem. To view Amazon Web Services CLI example commands that use these keys, see Creating OpsItems manually in the Amazon Web Services Systems Manager User Guide.

category: Option<String>

An OpsItem category. Category options include: Availability, Cost, Performance, Recovery, Security.

severity: Option<String>

The severity of the OpsItem. Severity options range from 1 to 4.

actual_start_time: Option<Instant>

The time a runbook workflow started. Currently reported only for the OpsItem type /aws/changerequest.

actual_end_time: Option<Instant>

The time a runbook workflow ended. Currently reported only for the OpsItem type /aws/changerequest.

planned_start_time: Option<Instant>

The time specified in a change request for a runbook workflow to start. Currently supported only for the OpsItem type /aws/changerequest.

planned_end_time: Option<Instant>

The time specified in a change request for a runbook workflow to end. Currently supported only for the OpsItem type /aws/changerequest.

Implementations

impl OpsItem

pub fn builder() -> Builder

Creates a new builder-style object to manufacture OpsItem