Struct aws_sdk_ssm::output::get_command_invocation_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for GetCommandInvocationOutput

Implementations

impl Builder

pub fn command_id(self, input: impl Into<String>) -> Self

The parent command ID of the invocation plugin.

pub fn set_command_id(self, input: Option<String>) -> Self

pub fn instance_id(self, input: impl Into<String>) -> Self

The ID of the managed instance targeted by the command. A managed instance can be an EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

pub fn set_instance_id(self, input: Option<String>) -> Self

pub fn comment(self, input: impl Into<String>) -> Self

The comment text for the command.

pub fn set_comment(self, input: Option<String>) -> Self

pub fn document_name(self, input: impl Into<String>) -> Self

The name of the document that was run. For example, AWS-RunShellScript.

pub fn set_document_name(self, input: Option<String>) -> Self

pub fn document_version(self, input: impl Into<String>) -> Self

The Systems Manager document (SSM document) version used in the request.

pub fn set_document_version(self, input: Option<String>) -> Self

pub fn plugin_name(self, input: impl Into<String>) -> Self

The name of the plugin, or step name, for which details are reported. For example, aws:RunShellScript is a plugin.

pub fn set_plugin_name(self, input: Option<String>) -> Self

pub fn response_code(self, input: i32) -> Self

The error level response code for the plugin script. If the response code is -1, then the command hasn't started running on the instance, or it wasn't received by the instance.

pub fn set_response_code(self, input: Option<i32>) -> Self

pub fn execution_start_date_time(self, input: impl Into<String>) -> Self

The date and time the plugin started running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedBefore filter.

aws ssm list-commands --filters key=InvokedBefore,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

pub fn set_execution_start_date_time(self, input: Option<String>) -> Self

pub fn execution_elapsed_time(self, input: impl Into<String>) -> Self

Duration since ExecutionStartDateTime.

pub fn set_execution_elapsed_time(self, input: Option<String>) -> Self

pub fn execution_end_date_time(self, input: impl Into<String>) -> Self

The date and time the plugin finished running. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample Amazon Web Services CLI command uses the InvokedAfter filter.

aws ssm list-commands --filters key=InvokedAfter,value=2017-06-07T00:00:00Z

If the plugin hasn't started to run, the string is empty.

pub fn set_execution_end_date_time(self, input: Option<String>) -> Self

pub fn status(self, input: CommandInvocationStatus) -> Self

The status of this invocation plugin. This status can be different than StatusDetails.

pub fn set_status(self, input: Option<CommandInvocationStatus>) -> Self

pub fn status_details(self, input: impl Into<String>) -> Self

A detailed status of the command execution for an invocation. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Understanding command statuses in the Amazon Web Services Systems Manager User Guide. StatusDetails can be one of the following values:

Pending: The command hasn't been sent to the instance.

In Progress: The command has been sent to the instance but hasn't reached a terminal state.

Delayed: The system attempted to send the command to the target, but the target wasn't available. The instance might not be available because of network issues, because the instance was stopped, or for similar reasons. The system will try to send the command again.

Success: The command or plugin ran successfully. This is a terminal state.

Delivery Timed Out: The command wasn't delivered to the instance before the delivery timeout expired. Delivery timeouts don't count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Execution Timed Out: The command started to run on the instance, but the execution wasn't complete before the timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state.

Failed: The command wasn't run successfully on the instance. For a plugin, this indicates that the result code wasn't zero. For a command invocation, this indicates that the result code for one or more plugins wasn't zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state.

Canceled: The command was terminated before it was completed. This is a terminal state.

Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state.

Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

pub fn set_status_details(self, input: Option<String>) -> Self

pub fn standard_output_content(self, input: impl Into<String>) -> Self

The first 24,000 characters written by the plugin to stdout. If the command hasn't finished running, if ExecutionStatus is neither Succeeded nor Failed, then this string is empty.

pub fn set_standard_output_content(self, input: Option<String>) -> Self

pub fn standard_output_url(self, input: impl Into<String>) -> Self

The URL for the complete text written by the plugin to stdout in Amazon Simple Storage Service (Amazon S3). If an S3 bucket wasn't specified, then this string is empty.

pub fn set_standard_output_url(self, input: Option<String>) -> Self

pub fn standard_error_content(self, input: impl Into<String>) -> Self

The first 8,000 characters written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

pub fn set_standard_error_content(self, input: Option<String>) -> Self

pub fn standard_error_url(self, input: impl Into<String>) -> Self

The URL for the complete text written by the plugin to stderr. If the command hasn't finished running, then this string is empty.

pub fn set_standard_error_url(self, input: Option<String>) -> Self

pub fn cloud_watch_output_config(self, input: CloudWatchOutputConfig) -> Self

Amazon CloudWatch Logs information where Systems Manager sent the command output.

pub fn set_cloud_watch_output_config(
self,
input: Option<CloudWatchOutputConfig>
) -> Self

pub fn build(self) -> GetCommandInvocationOutput

Consumes the builder and constructs a GetCommandInvocationOutput