Struct aws_sdk_ssm::output::get_document_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for GetDocumentOutput

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the Systems Manager document.

pub fn set_name(self, input: Option<String>) -> Self

pub fn created_date(self, input: Instant) -> Self

The date the Systems Manager document was created.

pub fn set_created_date(self, input: Option<Instant>) -> Self

pub fn display_name(self, input: impl Into<String>) -> Self

The friendly name of the Systems Manager document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

pub fn set_display_name(self, input: Option<String>) -> Self

pub fn version_name(self, input: impl Into<String>) -> Self

The version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and cannot be changed.

pub fn set_version_name(self, input: Option<String>) -> Self

pub fn document_version(self, input: impl Into<String>) -> Self

The document version.

pub fn set_document_version(self, input: Option<String>) -> Self

pub fn status(self, input: DocumentStatus) -> Self

The status of the Systems Manager document, such as Creating, Active, Updating, Failed, and Deleting.

pub fn set_status(self, input: Option<DocumentStatus>) -> Self

pub fn status_information(self, input: impl Into<String>) -> Self

A message returned by AWS Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket does not exist. Verify that the URL of the S3 bucket is correct."

pub fn set_status_information(self, input: Option<String>) -> Self

pub fn content(self, input: impl Into<String>) -> Self

The contents of the Systems Manager document.

pub fn set_content(self, input: Option<String>) -> Self

pub fn document_type(self, input: DocumentType) -> Self

The document type.

pub fn set_document_type(self, input: Option<DocumentType>) -> Self

pub fn document_format(self, input: DocumentFormat) -> Self

The document format, either JSON or YAML.

pub fn set_document_format(self, input: Option<DocumentFormat>) -> Self

pub fn requires(self, input: impl Into<DocumentRequires>) -> Self

pub fn set_requires(self, input: Option<Vec<DocumentRequires>>) -> Self

pub fn attachments_content(self, input: impl Into<AttachmentContent>) -> Self

pub fn set_attachments_content(
self,
input: Option<Vec<AttachmentContent>>
) -> Self

pub fn review_status(self, input: ReviewStatus) -> Self

The current review status of a new custom Systems Manager document (SSM document) created by a member of your organization, or of the latest version of an existing SSM document.

Only one version of an SSM document can be in the APPROVED state at a time. When a new version is approved, the status of the previous version changes to REJECTED.

Only one version of an SSM document can be in review, or PENDING, at a time.

pub fn set_review_status(self, input: Option<ReviewStatus>) -> Self

pub fn build(self) -> GetDocumentOutput

Consumes the builder and constructs a GetDocumentOutput