Struct aws_sdk_ssm::output::get_maintenance_window_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for GetMaintenanceWindowOutput

Implementations

impl Builder

pub fn window_id(self, inp: impl Into<String>) -> Self

The ID of the created maintenance window.

pub fn set_window_id(self, inp: Option<String>) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

The name of the maintenance window.

pub fn set_name(self, inp: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

The description of the maintenance window.

pub fn set_description(self, inp: Option<String>) -> Self

pub fn start_date(self, inp: impl Into<String>) -> Self

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active. The maintenance window will not run before this specified time.

pub fn set_start_date(self, inp: Option<String>) -> Self

pub fn end_date(self, inp: impl Into<String>) -> Self

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive. The maintenance window will not run after this specified time.

pub fn set_end_date(self, inp: Option<String>) -> Self

pub fn schedule(self, inp: impl Into<String>) -> Self

The schedule of the maintenance window in the form of a cron or rate expression.

pub fn set_schedule(self, inp: Option<String>) -> Self

pub fn schedule_timezone(self, inp: impl Into<String>) -> Self

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los_Angeles", "UTC", or "Asia/Seoul". For more information, see the Time Zone Database on the IANA website.

pub fn set_schedule_timezone(self, inp: Option<String>) -> Self

pub fn schedule_offset(self, inp: i32) -> Self

The number of days to wait to run a maintenance window after the scheduled CRON expression date and time.

pub fn set_schedule_offset(self, inp: Option<i32>) -> Self

pub fn next_execution_time(self, inp: impl Into<String>) -> Self

The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

pub fn set_next_execution_time(self, inp: Option<String>) -> Self

pub fn duration(self, inp: i32) -> Self

The duration of the maintenance window in hours.

pub fn set_duration(self, inp: i32) -> Self

pub fn cutoff(self, inp: i32) -> Self

The number of hours before the end of the maintenance window that Systems Manager stops scheduling new tasks for execution.

pub fn set_cutoff(self, inp: i32) -> Self

pub fn allow_unassociated_targets(self, inp: bool) -> Self

Whether targets must be registered with the maintenance window before tasks can be defined for those targets.

pub fn set_allow_unassociated_targets(self, inp: bool) -> Self

pub fn enabled(self, inp: bool) -> Self

Indicates whether the maintenance window is enabled.

pub fn set_enabled(self, inp: bool) -> Self

pub fn created_date(self, inp: Instant) -> Self

The date the maintenance window was created.

pub fn set_created_date(self, inp: Option<Instant>) -> Self

pub fn modified_date(self, inp: Instant) -> Self

The date the maintenance window was last modified.

pub fn set_modified_date(self, inp: Option<Instant>) -> Self

pub fn build(self) -> GetMaintenanceWindowOutput

Consumes the builder and constructs a GetMaintenanceWindowOutput