Struct aws_sdk_ssm::output::update_maintenance_window_task_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UpdateMaintenanceWindowTaskOutput

Implementations

impl Builder

pub fn window_id(self, input: impl Into<String>) -> Self

The ID of the maintenance window that was updated.

pub fn set_window_id(self, input: Option<String>) -> Self

pub fn window_task_id(self, input: impl Into<String>) -> Self

The task ID of the maintenance window that was updated.

pub fn set_window_task_id(self, input: Option<String>) -> Self

pub fn targets(self, input: impl Into<Target>) -> Self

pub fn set_targets(self, input: Option<Vec<Target>>) -> Self

pub fn task_arn(self, input: impl Into<String>) -> Self

The updated task ARN value.

pub fn set_task_arn(self, input: Option<String>) -> Self

pub fn service_role_arn(self, input: impl Into<String>) -> Self

The ARN of the IAM service role to use to publish Amazon Simple Notification Service (Amazon SNS) notifications for maintenance window Run Command tasks.

pub fn set_service_role_arn(self, input: Option<String>) -> Self

pub fn task_parameters(
self,
k: impl Into<String>,
v: impl Into<MaintenanceWindowTaskParameterValueExpression>
) -> Self

pub fn set_task_parameters(
self,
input: Option<HashMap<String, MaintenanceWindowTaskParameterValueExpression>>
) -> Self

pub fn task_invocation_parameters(
self,
input: MaintenanceWindowTaskInvocationParameters
) -> Self

The updated parameter values.

pub fn set_task_invocation_parameters(
self,
input: Option<MaintenanceWindowTaskInvocationParameters>
) -> Self

pub fn priority(self, input: i32) -> Self

The updated priority value.

pub fn set_priority(self, input: Option<i32>) -> Self

pub fn max_concurrency(self, input: impl Into<String>) -> Self

The updated MaxConcurrency value.

pub fn set_max_concurrency(self, input: Option<String>) -> Self

pub fn max_errors(self, input: impl Into<String>) -> Self

The updated MaxErrors value.

pub fn set_max_errors(self, input: Option<String>) -> Self

pub fn logging_info(self, input: LoggingInfo) -> Self

The updated logging information in Amazon S3.

LoggingInfo has been deprecated. To specify an S3 bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

pub fn set_logging_info(self, input: Option<LoggingInfo>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The updated task name.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The updated task description.

pub fn set_description(self, input: Option<String>) -> Self

pub fn build(self) -> UpdateMaintenanceWindowTaskOutput

Consumes the builder and constructs a UpdateMaintenanceWindowTaskOutput