Module aws_sdk_ssm::error

Modules

already_exists_error
See AlreadyExistsError

associated_instances
See AssociatedInstances

association_already_exists
See AssociationAlreadyExists

association_does_not_exist
See AssociationDoesNotExist

association_execution_does_not_exist
See AssociationExecutionDoesNotExist

association_limit_exceeded
See AssociationLimitExceeded

association_version_limit_exceeded
See AssociationVersionLimitExceeded

automation_definition_not_approved_error
See AutomationDefinitionNotApprovedError

automation_definition_not_found_error
See AutomationDefinitionNotFoundError

automation_definition_version_not_found_error
See AutomationDefinitionVersionNotFoundError

automation_execution_limit_exceeded_error
See AutomationExecutionLimitExceededError

automation_execution_not_found_error
See AutomationExecutionNotFoundError

automation_step_not_found_error
See AutomationStepNotFoundError

compliance_type_count_limit_exceeded_error
See ComplianceTypeCountLimitExceededError

custom_schema_count_limit_exceeded_error
See CustomSchemaCountLimitExceededError

document_already_exists
See DocumentAlreadyExists

document_limit_exceeded
See DocumentLimitExceeded

document_permission_limit
See DocumentPermissionLimit

document_version_limit_exceeded
See DocumentVersionLimitExceeded

does_not_exist_error
See DoesNotExistError

duplicate_document_content
See DuplicateDocumentContent

duplicate_document_version_name
See DuplicateDocumentVersionName

duplicate_instance_id
See DuplicateInstanceId

feature_not_available_error
See FeatureNotAvailableError

hierarchy_level_limit_exceeded_error
See HierarchyLevelLimitExceededError

hierarchy_type_mismatch_error
See HierarchyTypeMismatchError

idempotent_parameter_mismatch
See IdempotentParameterMismatch

incompatible_policy_error
See IncompatiblePolicyError

internal_server_error
See InternalServerError

invalid_activation
See InvalidActivation

invalid_activation_id
See InvalidActivationId

invalid_aggregator_error
See InvalidAggregatorError

invalid_allowed_pattern_error
See InvalidAllowedPatternError

invalid_association
See InvalidAssociation

invalid_association_version
See InvalidAssociationVersion

invalid_automation_execution_parameters_error
See InvalidAutomationExecutionParametersError

invalid_automation_signal_error
See InvalidAutomationSignalError

invalid_automation_status_update_error
See InvalidAutomationStatusUpdateError

invalid_command_id
See InvalidCommandId

invalid_delete_inventory_parameters_error
See InvalidDeleteInventoryParametersError

invalid_deletion_id_error
See InvalidDeletionIdError

invalid_document
See InvalidDocument

invalid_document_content
See InvalidDocumentContent

invalid_document_operation
See InvalidDocumentOperation

invalid_document_schema_version
See InvalidDocumentSchemaVersion

invalid_document_type
See InvalidDocumentType

invalid_document_version
See InvalidDocumentVersion

invalid_filter
See InvalidFilter

invalid_filter_key
See InvalidFilterKey

invalid_filter_option
See InvalidFilterOption

invalid_filter_value
See InvalidFilterValue

invalid_instance_id
See InvalidInstanceId

invalid_instance_information_filter_value
See InvalidInstanceInformationFilterValue

invalid_inventory_group_error
See InvalidInventoryGroupError

invalid_inventory_item_context_error
See InvalidInventoryItemContextError

invalid_inventory_request_error
See InvalidInventoryRequestError

invalid_item_content_error
See InvalidItemContentError

invalid_key_id
See InvalidKeyId

invalid_next_token
See InvalidNextToken

invalid_notification_config
See InvalidNotificationConfig

invalid_option_error
See InvalidOptionError

invalid_output_folder
See InvalidOutputFolder

invalid_output_location
See InvalidOutputLocation

invalid_parameters
See InvalidParameters

invalid_permission_type
See InvalidPermissionType

invalid_plugin_name
See InvalidPluginName

invalid_policy_attribute_error
See InvalidPolicyAttributeError

invalid_policy_type_error
See InvalidPolicyTypeError

invalid_resource_id
See InvalidResourceId

invalid_resource_type
See InvalidResourceType

invalid_result_attribute_error
See InvalidResultAttributeError

invalid_role
See InvalidRole

invalid_schedule
See InvalidSchedule

invalid_target
See InvalidTarget

invalid_type_name_error
See InvalidTypeNameError

invalid_update
See InvalidUpdate

invocation_does_not_exist
See InvocationDoesNotExist

item_content_mismatch_error
See ItemContentMismatchError

item_size_limit_exceeded_error
See ItemSizeLimitExceededError

max_document_size_exceeded
See MaxDocumentSizeExceeded

ops_item_already_exists_error
See OpsItemAlreadyExistsError

ops_item_invalid_parameter_error
See OpsItemInvalidParameterError

ops_item_limit_exceeded_error
See OpsItemLimitExceededError

ops_item_not_found_error
See OpsItemNotFoundError

ops_item_related_item_already_exists_error
See OpsItemRelatedItemAlreadyExistsError

ops_item_related_item_association_not_found_error
See OpsItemRelatedItemAssociationNotFoundError

ops_metadata_already_exists_error
See OpsMetadataAlreadyExistsError

ops_metadata_invalid_argument_error
See OpsMetadataInvalidArgumentError

ops_metadata_key_limit_exceeded_error
See OpsMetadataKeyLimitExceededError

ops_metadata_limit_exceeded_error
See OpsMetadataLimitExceededError

ops_metadata_not_found_error
See OpsMetadataNotFoundError

ops_metadata_too_many_updates_error
See OpsMetadataTooManyUpdatesError

parameter_already_exists
See ParameterAlreadyExists

parameter_limit_exceeded
See ParameterLimitExceeded

parameter_max_version_limit_exceeded
See ParameterMaxVersionLimitExceeded

parameter_not_found
See ParameterNotFound

parameter_pattern_mismatch_error
See ParameterPatternMismatchError

parameter_version_label_limit_exceeded
See ParameterVersionLabelLimitExceeded

parameter_version_not_found
See ParameterVersionNotFound

policies_limit_exceeded_error
See PoliciesLimitExceededError

resource_data_sync_already_exists_error
See ResourceDataSyncAlreadyExistsError

resource_data_sync_conflict_error
See ResourceDataSyncConflictError

resource_data_sync_count_exceeded_error
See ResourceDataSyncCountExceededError

resource_data_sync_invalid_configuration_error
See ResourceDataSyncInvalidConfigurationError

resource_data_sync_not_found_error
See ResourceDataSyncNotFoundError

resource_in_use_error
See ResourceInUseError

resource_limit_exceeded_error
See ResourceLimitExceededError

service_setting_not_found
See ServiceSettingNotFound

status_unchanged
See StatusUnchanged

sub_type_count_limit_exceeded_error
See SubTypeCountLimitExceededError

target_in_use_error
See TargetInUseError

target_not_connected
See TargetNotConnected

too_many_tags_error
See TooManyTagsError

too_many_updates
See TooManyUpdates

total_size_limit_exceeded_error
See TotalSizeLimitExceededError

unsupported_calendar_error
See UnsupportedCalendarError

unsupported_feature_required_error
See UnsupportedFeatureRequiredError

unsupported_inventory_item_context_error
See UnsupportedInventoryItemContextError

unsupported_inventory_schema_version_error
See UnsupportedInventorySchemaVersionError

unsupported_operating_system
See UnsupportedOperatingSystem

unsupported_parameter_type
See UnsupportedParameterType

unsupported_platform_type
See UnsupportedPlatformType

Structs

AddTagsToResourceError

AlreadyExistsError
Error returned if an attempt is made to register a patch group with a patch baseline that is

AssociateOpsItemRelatedItemError

AssociatedInstances
You must disassociate a document from all instances before you can delete it.

AssociationAlreadyExists
The specified association already exists.

AssociationDoesNotExist
The specified association does not exist.

AssociationExecutionDoesNotExist
The specified execution ID does not exist. Verify the ID number and try again.

AssociationLimitExceeded
You can have at most 2,000 active associations.

AssociationVersionLimitExceeded
You have reached the maximum number versions allowed for an association. Each association

AutomationDefinitionNotApprovedError
Indicates that the Change Manager change template used in the change request was rejected or is

AutomationDefinitionNotFoundError
An Automation document with the specified name could not be found.

AutomationDefinitionVersionNotFoundError
An Automation document with the specified name and version could not be found.

AutomationExecutionLimitExceededError
The number of simultaneously running Automation executions exceeded the allowable

AutomationExecutionNotFoundError
There is no automation execution information for the requested automation execution

AutomationStepNotFoundError
The specified step name and execution ID don't exist. Verify the information and try

CancelCommandError

CancelMaintenanceWindowExecutionError

ComplianceTypeCountLimitExceededError
You specified too many custom compliance types. You can specify a maximum of 10 different

CreateActivationError

CreateAssociationBatchError

CreateAssociationError

CreateDocumentError

CreateMaintenanceWindowError

CreateOpsItemError

CreateOpsMetadataError

CreatePatchBaselineError

CreateResourceDataSyncError

CustomSchemaCountLimitExceededError
You have exceeded the limit for custom schemas. Delete one or more custom schemas and try

DeleteActivationError

DeleteAssociationError

DeleteDocumentError

DeleteInventoryError

DeleteMaintenanceWindowError

DeleteOpsMetadataError

DeleteParameterError

DeleteParametersError

DeletePatchBaselineError

DeleteResourceDataSyncError

DeregisterManagedInstanceError

DeregisterPatchBaselineForPatchGroupError

DeregisterTargetFromMaintenanceWindowError

DeregisterTaskFromMaintenanceWindowError

DescribeActivationsError

DescribeAssociationError

DescribeAssociationExecutionTargetsError

DescribeAssociationExecutionsError

DescribeAutomationExecutionsError

DescribeAutomationStepExecutionsError

DescribeAvailablePatchesError

DescribeDocumentError

DescribeDocumentPermissionError

DescribeEffectiveInstanceAssociationsError

DescribeEffectivePatchesForPatchBaselineError

DescribeInstanceAssociationsStatusError

DescribeInstanceInformationError

DescribeInstancePatchStatesError

DescribeInstancePatchStatesForPatchGroupError

DescribeInstancePatchesError

DescribeInventoryDeletionsError

DescribeMaintenanceWindowExecutionTaskInvocationsError

DescribeMaintenanceWindowExecutionTasksError

DescribeMaintenanceWindowExecutionsError

DescribeMaintenanceWindowScheduleError

DescribeMaintenanceWindowTargetsError

DescribeMaintenanceWindowTasksError

DescribeMaintenanceWindowsError

DescribeMaintenanceWindowsForTargetError

DescribeOpsItemsError

DescribeParametersError

DescribePatchBaselinesError

DescribePatchGroupStateError

DescribePatchGroupsError

DescribePatchPropertiesError

DescribeSessionsError

DisassociateOpsItemRelatedItemError

DocumentAlreadyExists
The specified document already exists.

DocumentLimitExceeded
You can have at most 500 active Systems Manager documents.

DocumentPermissionLimit
The document cannot be shared with more AWS user accounts. You can share a document with a

DocumentVersionLimitExceeded
The document has too many versions. Delete one or more document versions and try

DoesNotExistError
Error returned when the ID specified for a resource, such as a maintenance window or Patch

DuplicateDocumentContent
The content of the association document matches another document. Change the content of the

DuplicateDocumentVersionName
The version name has already been used in this document. Specify a different version name,

DuplicateInstanceId
You cannot specify an instance ID in more than one association.

FeatureNotAvailableError
You attempted to register a LAMBDA or STEP_FUNCTIONS task in a region where the

GetAutomationExecutionError

GetCalendarStateError

GetCommandInvocationError

GetConnectionStatusError

GetDefaultPatchBaselineError

GetDeployablePatchSnapshotForInstanceError

GetDocumentError

GetInventoryError

GetInventorySchemaError

GetMaintenanceWindowError

GetMaintenanceWindowExecutionError

GetMaintenanceWindowExecutionTaskError

GetMaintenanceWindowExecutionTaskInvocationError

GetMaintenanceWindowTaskError

GetOpsItemError

GetOpsMetadataError

GetOpsSummaryError

GetParameterError

GetParameterHistoryError

GetParametersByPathError

GetParametersError

GetPatchBaselineError

GetPatchBaselineForPatchGroupError

GetServiceSettingError

HierarchyLevelLimitExceededError
A hierarchy can have a maximum of 15 levels. For more information, see Requirements and

HierarchyTypeMismatchError
Parameter Store does not support changing a parameter type in a hierarchy. For example, you

IdempotentParameterMismatch
Error returned when an idempotent operation is retried and the parameters don't match the

IncompatiblePolicyError
There is a conflict in the policies specified for this parameter. You can't, for example,

InternalServerError
An error occurred on the server side.

InvalidActivation
The activation is not valid. The activation might have been deleted, or the ActivationId and

InvalidActivationId
The activation ID is not valid. Verify the you entered the correct ActivationId or

InvalidAggregatorError
The specified aggregator is not valid for inventory groups. Verify that the aggregator uses

InvalidAllowedPatternError
The request does not meet the regular expression requirement.

InvalidAssociation
The association is not valid or does not exist.

InvalidAssociationVersion
The version you specified is not valid. Use ListAssociationVersions to view all versions of

InvalidAutomationExecutionParametersError
The supplied parameters for invoking the specified Automation document are incorrect. For

InvalidAutomationSignalError
The signal is not valid for the current Automation execution.

InvalidAutomationStatusUpdateError
The specified update status operation is not valid.

InvalidCommandId

InvalidDeleteInventoryParametersError
One or more of the parameters specified for the delete operation is not valid. Verify all

InvalidDeletionIdError
The ID specified for the delete operation does not exist or is not valid. Verify the ID and

InvalidDocument
The specified document does not exist.

InvalidDocumentContent
The content for the document is not valid.

InvalidDocumentOperation
You attempted to delete a document while it is still shared. You must stop sharing the

InvalidDocumentSchemaVersion
The version of the document schema is not supported.

InvalidDocumentType
The document type is not valid. Valid document types are described in the

InvalidDocumentVersion
The document version is not valid or does not exist.

InvalidFilter
The filter name is not valid. Verify the you entered the correct name and try again.

InvalidFilterKey
The specified key is not valid.

InvalidFilterOption
The specified filter option is not valid. Valid options are Equals and BeginsWith. For Path

InvalidFilterValue
The filter value is not valid. Verify the value and try again.

InvalidInstanceId
The following problems can cause this exception:

InvalidInstanceInformationFilterValue
The specified filter value is not valid.

InvalidInventoryGroupError
The specified inventory group is not valid.

InvalidInventoryItemContextError
You specified invalid keys or values in the Context attribute for

InvalidInventoryRequestError
The request is not valid.

InvalidItemContentError
One or more content items is not valid.

InvalidKeyId
The query key ID is not valid.

InvalidNextToken
The specified token is not valid.

InvalidNotificationConfig
One or more configuration items is not valid. Verify that a valid Amazon Resource Name (ARN)

InvalidOptionError
The delete inventory option specified is not valid. Verify the option and try again.

InvalidOutputFolder
The S3 bucket does not exist.

InvalidOutputLocation
The output location is not valid or does not exist.

InvalidParameters
You must specify values for all required parameters in the Systems Manager document. You can only

InvalidPermissionType
The permission type is not supported. Share is the only supported

InvalidPluginName
The plugin name is not valid.

InvalidPolicyAttributeError
A policy attribute or its value is invalid.

InvalidPolicyTypeError
The policy type is not supported. Parameter Store supports the following policy types:

InvalidResourceId
The resource ID is not valid. Verify that you entered the correct ID and try again.

InvalidResourceType
The resource type is not valid. For example, if you are attempting to tag an instance, the

InvalidResultAttributeError
The specified inventory item result attribute is not valid.

InvalidRole
The role name can't contain invalid characters. Also verify that you specified an IAM role

InvalidSchedule
The schedule is invalid. Verify your cron or rate expression and try again.

InvalidTarget
The target is not valid or does not exist. It might not be configured for Systems Manager or you might

InvalidTypeNameError
The parameter type name is not valid.

InvalidUpdate
The update is not valid.

InvocationDoesNotExist
The command ID and instance ID you specified did not match any invocations. Verify the

ItemContentMismatchError
The inventory item has invalid content.

ItemSizeLimitExceededError
The inventory item size has exceeded the size limit.

LabelParameterVersionError

ListAssociationVersionsError

ListAssociationsError

ListCommandInvocationsError

ListCommandsError

ListComplianceItemsError

ListComplianceSummariesError

ListDocumentMetadataHistoryError

ListDocumentVersionsError

ListDocumentsError

ListInventoryEntriesError

ListOpsItemEventsError

ListOpsItemRelatedItemsError

ListOpsMetadataError

ListResourceComplianceSummariesError

ListResourceDataSyncError

ListTagsForResourceError

MaxDocumentSizeExceeded
The size limit of a document is 64 KB.

ModifyDocumentPermissionError

OpsItemAlreadyExistsError
The OpsItem already exists.

OpsItemInvalidParameterError
A specified parameter argument isn't valid. Verify the available arguments and try

OpsItemLimitExceededError
The request caused OpsItems to exceed one or more quotas. For information about OpsItem

OpsItemNotFoundError
The specified OpsItem ID doesn't exist. Verify the ID and try again.

OpsItemRelatedItemAlreadyExistsError
The Amazon Resource Name (ARN) is already associated with the OpsItem.

OpsItemRelatedItemAssociationNotFoundError
The association was not found using the parameters you specified in the call. Verify the

OpsMetadataAlreadyExistsError
An OpsMetadata object already exists for the selected resource.

OpsMetadataInvalidArgumentError
One of the arguments passed is invalid.

OpsMetadataKeyLimitExceededError
The OpsMetadata object exceeds the maximum number of OpsMetadata keys that you can assign to

OpsMetadataLimitExceededError
Your account reached the maximum number of OpsMetadata objects allowed by Application Manager. The

OpsMetadataNotFoundError
The OpsMetadata object does not exist.

OpsMetadataTooManyUpdatesError
The system is processing too many concurrent updates. Wait a few moments and try

ParameterAlreadyExists
The parameter already exists. You can't create duplicate parameters.

ParameterLimitExceeded
You have exceeded the number of parameters for this AWS account. Delete one or more

ParameterMaxVersionLimitExceeded
Parameter Store retains the 100 most recently created versions of a parameter. After this

ParameterNotFound
The parameter could not be found. Verify the name and try again.

ParameterPatternMismatchError
The parameter name is not valid.

ParameterVersionLabelLimitExceeded
A parameter version can have a maximum of ten labels.

ParameterVersionNotFound
The specified parameter version was not found. Verify the parameter name and version, and

PoliciesLimitExceededError
You specified more than the maximum number of allowed policies for the parameter. The

PutComplianceItemsError

PutInventoryError

PutParameterError

RegisterDefaultPatchBaselineError

RegisterPatchBaselineForPatchGroupError

RegisterTargetWithMaintenanceWindowError

RegisterTaskWithMaintenanceWindowError

RemoveTagsFromResourceError

ResetServiceSettingError

ResourceDataSyncAlreadyExistsError
A sync configuration with the same name already exists.

ResourceDataSyncConflictError
Another UpdateResourceDataSync request is being processed. Wait a few minutes

ResourceDataSyncCountExceededError
You have exceeded the allowed maximum sync configurations.

ResourceDataSyncInvalidConfigurationError
The specified sync configuration is invalid.

ResourceDataSyncNotFoundError
The specified sync name was not found.

ResourceInUseError
Error returned if an attempt is made to delete a patch baseline that is registered for a

ResourceLimitExceededError
Error returned when the caller has exceeded the default resource quotas. For example, too

ResumeSessionError

SendAutomationSignalError

SendCommandError

ServiceSettingNotFound
The specified service setting was not found. Either the service name or the setting has not

StartAssociationsOnceError

StartAutomationExecutionError

StartChangeRequestExecutionError

StartSessionError

StatusUnchanged
The updated status is the same as the current status.

StopAutomationExecutionError

SubTypeCountLimitExceededError
The sub-type count exceeded the limit for the inventory type.

TargetInUseError
You specified the Safe option for the DeregisterTargetFromMaintenanceWindow

TargetNotConnected
The specified target instance for the session is not fully configured for use with Session Manager.

TerminateSessionError

TooManyTagsError
The Targets parameter includes too many tags. Remove one or more tags and try

TooManyUpdates
There are concurrent updates for a resource that supports one update at a time.

TotalSizeLimitExceededError
The size of inventory data has exceeded the total size limit for the resource.

UnlabelParameterVersionError

UnsupportedCalendarError
The calendar entry contained in the specified Systems Manager document is not supported.

UnsupportedFeatureRequiredError
Microsoft application patching is only available on EC2 instances and advanced instances. To

UnsupportedInventoryItemContextError
The Context attribute that you specified for the InventoryItem is

UnsupportedInventorySchemaVersionError
Inventory item type schema version has to match supported versions in the service. Check

UnsupportedOperatingSystem
The operating systems you specified is not supported, or the operation is not supported for

UnsupportedParameterType
The parameter type is not supported.

UnsupportedPlatformType
The document does not support the platform type of the given instance ID(s). For example,

UpdateAssociationError

UpdateAssociationStatusError

UpdateDocumentDefaultVersionError

UpdateDocumentError

UpdateDocumentMetadataError

UpdateMaintenanceWindowError

UpdateMaintenanceWindowTargetError

UpdateMaintenanceWindowTaskError

UpdateManagedInstanceRoleError

UpdateOpsItemError

UpdateOpsMetadataError

UpdatePatchBaselineError

UpdateResourceDataSyncError

UpdateServiceSettingError

Enums

AddTagsToResourceErrorKind

AssociateOpsItemRelatedItemErrorKind

CancelCommandErrorKind

CancelMaintenanceWindowExecutionErrorKind

CreateActivationErrorKind

CreateAssociationBatchErrorKind

CreateAssociationErrorKind

CreateDocumentErrorKind

CreateMaintenanceWindowErrorKind

CreateOpsItemErrorKind

CreateOpsMetadataErrorKind

CreatePatchBaselineErrorKind

CreateResourceDataSyncErrorKind

DeleteActivationErrorKind

DeleteAssociationErrorKind

DeleteDocumentErrorKind

DeleteInventoryErrorKind

DeleteMaintenanceWindowErrorKind

DeleteOpsMetadataErrorKind

DeleteParameterErrorKind

DeleteParametersErrorKind

DeletePatchBaselineErrorKind

DeleteResourceDataSyncErrorKind

DeregisterManagedInstanceErrorKind

DeregisterPatchBaselineForPatchGroupErrorKind

DeregisterTargetFromMaintenanceWindowErrorKind

DeregisterTaskFromMaintenanceWindowErrorKind

DescribeActivationsErrorKind

DescribeAssociationErrorKind

DescribeAssociationExecutionTargetsErrorKind

DescribeAssociationExecutionsErrorKind

DescribeAutomationExecutionsErrorKind

DescribeAutomationStepExecutionsErrorKind

DescribeAvailablePatchesErrorKind

DescribeDocumentErrorKind

DescribeDocumentPermissionErrorKind

DescribeEffectiveInstanceAssociationsErrorKind

DescribeEffectivePatchesForPatchBaselineErrorKind

DescribeInstanceAssociationsStatusErrorKind

DescribeInstanceInformationErrorKind

DescribeInstancePatchStatesErrorKind

DescribeInstancePatchStatesForPatchGroupErrorKind

DescribeInstancePatchesErrorKind

DescribeInventoryDeletionsErrorKind

DescribeMaintenanceWindowExecutionTaskInvocationsErrorKind

DescribeMaintenanceWindowExecutionTasksErrorKind

DescribeMaintenanceWindowExecutionsErrorKind

DescribeMaintenanceWindowScheduleErrorKind

DescribeMaintenanceWindowTargetsErrorKind

DescribeMaintenanceWindowTasksErrorKind

DescribeMaintenanceWindowsErrorKind

DescribeMaintenanceWindowsForTargetErrorKind

DescribeOpsItemsErrorKind

DescribeParametersErrorKind

DescribePatchBaselinesErrorKind

DescribePatchGroupStateErrorKind

DescribePatchGroupsErrorKind

DescribePatchPropertiesErrorKind

DescribeSessionsErrorKind

DisassociateOpsItemRelatedItemErrorKind

GetAutomationExecutionErrorKind

GetCalendarStateErrorKind

GetCommandInvocationErrorKind

GetConnectionStatusErrorKind

GetDefaultPatchBaselineErrorKind

GetDeployablePatchSnapshotForInstanceErrorKind

GetDocumentErrorKind

GetInventoryErrorKind

GetInventorySchemaErrorKind

GetMaintenanceWindowErrorKind

GetMaintenanceWindowExecutionErrorKind

GetMaintenanceWindowExecutionTaskErrorKind

GetMaintenanceWindowExecutionTaskInvocationErrorKind

GetMaintenanceWindowTaskErrorKind

GetOpsItemErrorKind

GetOpsMetadataErrorKind

GetOpsSummaryErrorKind

GetParameterErrorKind

GetParameterHistoryErrorKind

GetParametersByPathErrorKind

GetParametersErrorKind

GetPatchBaselineErrorKind

GetPatchBaselineForPatchGroupErrorKind

GetServiceSettingErrorKind

LabelParameterVersionErrorKind

ListAssociationVersionsErrorKind

ListAssociationsErrorKind

ListCommandInvocationsErrorKind

ListCommandsErrorKind

ListComplianceItemsErrorKind

ListComplianceSummariesErrorKind

ListDocumentMetadataHistoryErrorKind

ListDocumentVersionsErrorKind

ListDocumentsErrorKind

ListInventoryEntriesErrorKind

ListOpsItemEventsErrorKind

ListOpsItemRelatedItemsErrorKind

ListOpsMetadataErrorKind

ListResourceComplianceSummariesErrorKind

ListResourceDataSyncErrorKind

ListTagsForResourceErrorKind

ModifyDocumentPermissionErrorKind

PutComplianceItemsErrorKind

PutInventoryErrorKind

PutParameterErrorKind

RegisterDefaultPatchBaselineErrorKind

RegisterPatchBaselineForPatchGroupErrorKind

RegisterTargetWithMaintenanceWindowErrorKind

RegisterTaskWithMaintenanceWindowErrorKind

RemoveTagsFromResourceErrorKind

ResetServiceSettingErrorKind

ResumeSessionErrorKind

SendAutomationSignalErrorKind

SendCommandErrorKind

StartAssociationsOnceErrorKind

StartAutomationExecutionErrorKind

StartChangeRequestExecutionErrorKind

StartSessionErrorKind

StopAutomationExecutionErrorKind

TerminateSessionErrorKind

UnlabelParameterVersionErrorKind

UpdateAssociationErrorKind

UpdateAssociationStatusErrorKind

UpdateDocumentDefaultVersionErrorKind

UpdateDocumentErrorKind

UpdateDocumentMetadataErrorKind

UpdateMaintenanceWindowErrorKind

UpdateMaintenanceWindowTargetErrorKind

UpdateMaintenanceWindowTaskErrorKind

UpdateManagedInstanceRoleErrorKind

UpdateOpsItemErrorKind

UpdateOpsMetadataErrorKind

UpdatePatchBaselineErrorKind

UpdateResourceDataSyncErrorKind

UpdateServiceSettingErrorKind