Struct aws_sdk_ssm::model::automation_execution::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AutomationExecution

Implementations

impl Builder

pub fn automation_execution_id(self, inp: impl Into<String>) -> Self

The execution ID.

pub fn set_automation_execution_id(self, inp: Option<String>) -> Self

pub fn document_name(self, inp: impl Into<String>) -> Self

The name of the Automation document used during the execution.

pub fn set_document_name(self, inp: Option<String>) -> Self

pub fn document_version(self, inp: impl Into<String>) -> Self

The version of the document to use during execution.

pub fn set_document_version(self, inp: Option<String>) -> Self

pub fn execution_start_time(self, inp: Instant) -> Self

The time the execution started.

pub fn set_execution_start_time(self, inp: Option<Instant>) -> Self

pub fn execution_end_time(self, inp: Instant) -> Self

The time the execution finished.

pub fn set_execution_end_time(self, inp: Option<Instant>) -> Self

pub fn automation_execution_status(self, inp: AutomationExecutionStatus) -> Self

The execution status of the Automation.

pub fn set_automation_execution_status(
self,
inp: Option<AutomationExecutionStatus>
) -> Self

pub fn step_executions(self, inp: impl Into<StepExecution>) -> Self

pub fn set_step_executions(self, inp: Option<Vec<StepExecution>>) -> Self

pub fn step_executions_truncated(self, inp: bool) -> Self

A boolean value that indicates if the response contains the full list of the Automation step executions. If true, use the DescribeAutomationStepExecutions API action to get the full list of step executions.

pub fn set_step_executions_truncated(self, inp: bool) -> Self

pub fn parameters(self, k: impl Into<String>, v: impl Into<Vec<String>>) -> Self

pub fn set_parameters(self, inp: Option<HashMap<String, Vec<String>>>) -> Self

pub fn outputs(self, k: impl Into<String>, v: impl Into<Vec<String>>) -> Self

pub fn set_outputs(self, inp: Option<HashMap<String, Vec<String>>>) -> Self

pub fn failure_message(self, inp: impl Into<String>) -> Self

A message describing why an execution has failed, if the status is set to Failed.

pub fn set_failure_message(self, inp: Option<String>) -> Self

pub fn mode(self, inp: ExecutionMode) -> Self

The automation execution mode.

pub fn set_mode(self, inp: Option<ExecutionMode>) -> Self

pub fn parent_automation_execution_id(self, inp: impl Into<String>) -> Self

The AutomationExecutionId of the parent automation.

pub fn set_parent_automation_execution_id(self, inp: Option<String>) -> Self

pub fn executed_by(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the user who ran the automation.

pub fn set_executed_by(self, inp: Option<String>) -> Self

pub fn current_step_name(self, inp: impl Into<String>) -> Self

The name of the step that is currently running.

pub fn set_current_step_name(self, inp: Option<String>) -> Self

pub fn current_action(self, inp: impl Into<String>) -> Self

The action of the step that is currently running.

pub fn set_current_action(self, inp: Option<String>) -> Self

pub fn target_parameter_name(self, inp: impl Into<String>) -> Self

The parameter name.

pub fn set_target_parameter_name(self, inp: Option<String>) -> Self

pub fn targets(self, inp: impl Into<Target>) -> Self

pub fn set_targets(self, inp: Option<Vec<Target>>) -> Self

pub fn target_maps(self, inp: impl Into<HashMap<String, Vec<String>>>) -> Self

pub fn set_target_maps(
self,
inp: Option<Vec<HashMap<String, Vec<String>>>>
) -> Self

pub fn resolved_targets(self, inp: ResolvedTargets) -> Self

A list of resolved targets in the rate control execution.

pub fn set_resolved_targets(self, inp: Option<ResolvedTargets>) -> Self

pub fn max_concurrency(self, inp: impl Into<String>) -> Self

The MaxConcurrency value specified by the user when the execution started.

pub fn set_max_concurrency(self, inp: Option<String>) -> Self

pub fn max_errors(self, inp: impl Into<String>) -> Self

The MaxErrors value specified by the user when the execution started.

pub fn set_max_errors(self, inp: Option<String>) -> Self

pub fn target(self, inp: impl Into<String>) -> Self

The target of the execution.

pub fn set_target(self, inp: Option<String>) -> Self

pub fn target_locations(self, inp: impl Into<TargetLocation>) -> Self

pub fn set_target_locations(self, inp: Option<Vec<TargetLocation>>) -> Self

pub fn progress_counters(self, inp: ProgressCounters) -> Self

An aggregate of step execution statuses displayed in the AWS Console for a multi-Region and multi-account Automation execution.

pub fn set_progress_counters(self, inp: Option<ProgressCounters>) -> Self

pub fn automation_subtype(self, inp: AutomationSubtype) -> Self

The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

pub fn set_automation_subtype(self, inp: Option<AutomationSubtype>) -> Self

pub fn scheduled_time(self, inp: Instant) -> Self

The date and time the Automation operation is scheduled to start.

pub fn set_scheduled_time(self, inp: Option<Instant>) -> Self

pub fn runbooks(self, inp: impl Into<Runbook>) -> Self

pub fn set_runbooks(self, inp: Option<Vec<Runbook>>) -> Self

pub fn ops_item_id(self, inp: impl Into<String>) -> Self

The ID of an OpsItem that is created to represent a Change Manager change request.

pub fn set_ops_item_id(self, inp: Option<String>) -> Self

pub fn association_id(self, inp: impl Into<String>) -> Self

The ID of a State Manager association used in the Automation operation.

pub fn set_association_id(self, inp: Option<String>) -> Self

pub fn change_request_name(self, inp: impl Into<String>) -> Self

The name of the Change Manager change request.

pub fn set_change_request_name(self, inp: Option<String>) -> Self

pub fn build(self) -> AutomationExecution

Consumes the builder and constructs a AutomationExecution