Struct aws_sdk_ssm::model::patch::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Patch

Implementations

impl Builder

pub fn id(self, inp: impl Into<String>) -> Self

The ID of the patch. Applies to Windows patches only.

This ID is not the same as the Microsoft Knowledge Base ID.

pub fn set_id(self, inp: Option<String>) -> Self

pub fn release_date(self, inp: Instant) -> Self

The date the patch was released.

pub fn set_release_date(self, inp: Option<Instant>) -> Self

pub fn title(self, inp: impl Into<String>) -> Self

The title of the patch.

pub fn set_title(self, inp: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

The description of the patch.

pub fn set_description(self, inp: Option<String>) -> Self

pub fn content_url(self, inp: impl Into<String>) -> Self

The URL where more information can be obtained about the patch.

pub fn set_content_url(self, inp: Option<String>) -> Self

pub fn vendor(self, inp: impl Into<String>) -> Self

The name of the vendor providing the patch.

pub fn set_vendor(self, inp: Option<String>) -> Self

pub fn product_family(self, inp: impl Into<String>) -> Self

The product family the patch is applicable for. For example, Windows or Amazon Linux 2.

pub fn set_product_family(self, inp: Option<String>) -> Self

pub fn product(self, inp: impl Into<String>) -> Self

The specific product the patch is applicable for. For example, WindowsServer2016 or AmazonLinux2018.03.

pub fn set_product(self, inp: Option<String>) -> Self

pub fn classification(self, inp: impl Into<String>) -> Self

The classification of the patch. For example, SecurityUpdates, Updates, or CriticalUpdates.

pub fn set_classification(self, inp: Option<String>) -> Self

pub fn msrc_severity(self, inp: impl Into<String>) -> Self

The severity of the patch, such as Critical, Important, or Moderate. Applies to Windows patches only.

pub fn set_msrc_severity(self, inp: Option<String>) -> Self

pub fn kb_number(self, inp: impl Into<String>) -> Self

The Microsoft Knowledge Base ID of the patch. Applies to Windows patches only.

pub fn set_kb_number(self, inp: Option<String>) -> Self

pub fn msrc_number(self, inp: impl Into<String>) -> Self

The ID of the Microsoft Security Response Center (MSRC) bulletin the patch is related to. For example, MS14-045. Applies to Windows patches only.

pub fn set_msrc_number(self, inp: Option<String>) -> Self

pub fn language(self, inp: impl Into<String>) -> Self

The language of the patch if it's language-specific.

pub fn set_language(self, inp: Option<String>) -> Self

pub fn advisory_ids(self, inp: impl Into<String>) -> Self

pub fn set_advisory_ids(self, inp: Option<Vec<String>>) -> Self

pub fn bugzilla_ids(self, inp: impl Into<String>) -> Self

pub fn set_bugzilla_ids(self, inp: Option<Vec<String>>) -> Self

pub fn cve_ids(self, inp: impl Into<String>) -> Self

pub fn set_cve_ids(self, inp: Option<Vec<String>>) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

The name of the patch. Applies to Linux-based instances only.

pub fn set_name(self, inp: Option<String>) -> Self

pub fn epoch(self, inp: i32) -> Self

The epoch of the patch. For example in pkg-example-EE-20180914-2.2.amzn1.noarch, the epoch value is 20180914-2. Applies to Linux-based instances only.

pub fn set_epoch(self, inp: i32) -> Self

pub fn version(self, inp: impl Into<String>) -> Self

The version number of the patch. For example, in example-pkg-1.710.10-2.7.abcd.x86_64, the version number is indicated by -1. Applies to Linux-based instances only.

pub fn set_version(self, inp: Option<String>) -> Self

pub fn release(self, inp: impl Into<String>) -> Self

The particular release of a patch. For example, in pkg-example-EE-20180914-2.2.amzn1.noarch, the release is 2.amaz1. Applies to Linux-based instances only.

pub fn set_release(self, inp: Option<String>) -> Self

pub fn arch(self, inp: impl Into<String>) -> Self

The architecture of the patch. For example, in example-pkg-0.710.10-2.7.abcd.x86_64, the architecture is indicated by x86_64. Applies to Linux-based instances only.

pub fn set_arch(self, inp: Option<String>) -> Self

pub fn severity(self, inp: impl Into<String>) -> Self

The severity level of the patch. For example, CRITICAL or MODERATE.

pub fn set_severity(self, inp: Option<String>) -> Self

pub fn repository(self, inp: impl Into<String>) -> Self

The source patch repository for the operating system and version, such as trusty-security for Ubuntu Server 14.04 LTE and focal-security for Ubuntu Server 20.04 LTE. Applies to Linux-based instances only.

pub fn set_repository(self, inp: Option<String>) -> Self

pub fn build(self) -> Patch

Consumes the builder and constructs a Patch