Module aws_sdk_ssm::operation

Structs

AddTagsToResource
Adds or overwrites one or more tags for the specified resource. Tags are metadata that you

AssociateOpsItemRelatedItem
Associates a related resource to a Systems Manager OpsCenter OpsItem. For example, you can

CancelCommand
Attempts to cancel the command specified by the Command ID. There is no guarantee that the

CancelMaintenanceWindowExecution
Stops a maintenance window execution that is already in progress and cancels any tasks in

CreateActivation
Generates an activation code and activation ID you can use to register your on-premises

CreateAssociation
A State Manager association defines the state that you want to maintain on your instances.

CreateAssociationBatch
Associates the specified Systems Manager document with the specified instances or targets.

CreateDocument
Creates a Systems Manager (SSM) document. An SSM document defines the actions that Systems Manager performs on

CreateMaintenanceWindow
Creates a new maintenance window.

CreateOpsItem
Creates a new OpsItem. You must have permission in AWS Identity and Access Management (IAM)

CreateOpsMetadata
If you create a new application in Application Manager, Systems Manager calls this API action to specify

CreatePatchBaseline
Creates a patch baseline.

CreateResourceDataSync
A resource data sync helps you view data from multiple sources in a single location. Systems

DeleteActivation
Deletes an activation. You are not required to delete an activation. If you delete an

DeleteAssociation
Disassociates the specified Systems Manager document from the specified instance.

DeleteDocument
Deletes the Systems Manager document and all instance associations to the document.

DeleteInventory
Delete a custom inventory type or the data associated with a custom Inventory type. Deleting

DeleteMaintenanceWindow
Deletes a maintenance window.

DeleteOpsMetadata
Delete OpsMetadata related to an application.

DeleteParameter
Delete a parameter from the system.

DeleteParameters
Delete a list of parameters.

DeletePatchBaseline
Deletes a patch baseline.

DeleteResourceDataSync
Deletes a Resource Data Sync configuration. After the configuration is deleted, changes to

DeregisterManagedInstance
Removes the server or virtual machine from the list of registered servers. You can

DeregisterPatchBaselineForPatchGroup
Removes a patch group from a patch baseline.

DeregisterTargetFromMaintenanceWindow
Removes a target from a maintenance window.

DeregisterTaskFromMaintenanceWindow
Removes a task from a maintenance window.

DescribeActivations
Describes details about the activation, such as the date and time the activation was

DescribeAssociation
Describes the association for the specified target or instance. If you created the

DescribeAssociationExecutionTargets
Use this API action to view information about a specific execution of a specific

DescribeAssociationExecutions
Use this API action to view all executions for a specific association ID.

DescribeAutomationExecutions
Provides details about all active and terminated Automation executions.

DescribeAutomationStepExecutions
Information about all active and terminated step executions in an Automation

DescribeAvailablePatches
Lists all patches eligible to be included in a patch baseline.

DescribeDocument
Describes the specified Systems Manager document.

DescribeDocumentPermission
Describes the permissions for a Systems Manager document. If you created the document, you are the

DescribeEffectiveInstanceAssociations
All associations for the instance(s).

DescribeEffectivePatchesForPatchBaseline
Retrieves the current effective patches (the patch and the approval state) for the specified

DescribeInstanceAssociationsStatus
The status of the associations for the instance(s).

DescribeInstanceInformation
Describes one or more of your instances, including information about the operating system

DescribeInstancePatchStates
Retrieves the high-level patch state of one or more instances.

DescribeInstancePatchStatesForPatchGroup
Retrieves the high-level patch state for the instances in the specified patch group.

DescribeInstancePatches
Retrieves information about the patches on the specified instance and their state relative

DescribeInventoryDeletions
Describes a specific delete inventory operation.

DescribeMaintenanceWindowExecutionTaskInvocations
Retrieves the individual task executions (one per target) for a particular task run as part

DescribeMaintenanceWindowExecutionTasks
For a given maintenance window execution, lists the tasks that were run.

DescribeMaintenanceWindowExecutions
Lists the executions of a maintenance window. This includes information about when the

DescribeMaintenanceWindowSchedule
Retrieves information about upcoming executions of a maintenance window.

DescribeMaintenanceWindowTargets
Lists the targets registered with the maintenance window.

DescribeMaintenanceWindowTasks
Lists the tasks in a maintenance window.

DescribeMaintenanceWindows
Retrieves the maintenance windows in an AWS account.

DescribeMaintenanceWindowsForTarget
Retrieves information about the maintenance window targets or tasks that an instance is

DescribeOpsItems
Query a set of OpsItems. You must have permission in AWS Identity and Access Management

DescribeParameters
Get information about a parameter.

DescribePatchBaselines
Lists the patch baselines in your AWS account.

DescribePatchGroupState
Returns high-level aggregated patch compliance state for a patch group.

DescribePatchGroups
Lists all patch groups that have been registered with patch baselines.

DescribePatchProperties
Lists the properties of available patches organized by product, product family,

DescribeSessions
Retrieves a list of all active sessions (both connected and disconnected) or terminated

DisassociateOpsItemRelatedItem
Deletes the association between an OpsItem and a related resource. For example, this API

GetAutomationExecution
Get detailed information about a particular Automation execution.

GetCalendarState
Gets the state of the AWS Systems Manager Change Calendar at an optional, specified time. If you

GetCommandInvocation
Returns detailed information about command execution for an invocation or plugin.

GetConnectionStatus
Retrieves the Session Manager connection status for an instance to determine whether it is running and

GetDefaultPatchBaseline
Retrieves the default patch baseline. Note that Systems Manager supports creating multiple default

GetDeployablePatchSnapshotForInstance
Retrieves the current snapshot for the patch baseline the instance uses. This API is

GetDocument
Gets the contents of the specified Systems Manager document.

GetInventory
Query inventory information.

GetInventorySchema
Return a list of inventory type names for the account, or return a list of attribute names

GetMaintenanceWindow
Retrieves a maintenance window.

GetMaintenanceWindowExecution
Retrieves details about a specific a maintenance window execution.

GetMaintenanceWindowExecutionTask
Retrieves the details about a specific task run as part of a maintenance window

GetMaintenanceWindowExecutionTaskInvocation
Retrieves information about a specific task running on a specific target.

GetMaintenanceWindowTask
Lists the tasks in a maintenance window.

GetOpsItem
Get information about an OpsItem by using the ID. You must have permission in AWS Identity

GetOpsMetadata
View operational metadata related to an application in Application Manager.

GetOpsSummary
View a summary of OpsItems based on specified filters and aggregators.

GetParameter
Get information about a parameter by using the parameter name. Don't confuse this API action

GetParameterHistory
Retrieves the history of all changes to a parameter.

GetParameters
Get details of a parameter. Don't confuse this API action with the GetParameter API action.

GetParametersByPath
Retrieve information about one or more parameters in a specific hierarchy.

GetPatchBaseline
Retrieves information about a patch baseline.

GetPatchBaselineForPatchGroup
Retrieves the patch baseline that should be used for the specified patch group.

GetServiceSetting

LabelParameterVersion
A parameter label is a user-defined alias to help you manage different versions of a

ListAssociationVersions
Retrieves all versions of an association for a specific association ID.

ListAssociations
Returns all State Manager associations in the current AWS account and Region. You can limit

ListCommandInvocations
An invocation is copy of a command sent to a specific instance. A command can apply to one

ListCommands
Lists the commands requested by users of the AWS account.

ListComplianceItems
For a specified resource ID, this API action returns a list of compliance statuses for

ListComplianceSummaries
Returns a summary count of compliant and non-compliant resources for a compliance type. For

ListDocumentMetadataHistory
Information about approval reviews for a version of an SSM document.

ListDocumentVersions
List all versions for a document.

ListDocuments
Returns all Systems Manager (SSM) documents in the current AWS account and Region. You can limit the

ListInventoryEntries
A list of inventory items returned by the request.

ListOpsItemEvents
Returns a list of all OpsItem events in the current AWS account and Region. You can limit

ListOpsItemRelatedItems
Lists all related-item resources associated with an OpsItem.

ListOpsMetadata
Systems Manager calls this API action when displaying all Application Manager OpsMetadata objects or

ListResourceComplianceSummaries
Returns a resource-level summary count. The summary includes information about compliant and

ListResourceDataSync
Lists your resource data sync configurations. Includes information about the last time a

ListTagsForResource
Returns a list of the tags assigned to the specified resource.

ModifyDocumentPermission
Shares a Systems Manager document publicly or privately. If you share a document privately, you must

PutComplianceItems
Registers a compliance type and other compliance details on a designated resource. This

PutInventory
Bulk update custom inventory items on one more instance. The request adds an inventory item,

PutParameter
Add a parameter to the system.

RegisterDefaultPatchBaseline
Defines the default patch baseline for the relevant operating system.

RegisterPatchBaselineForPatchGroup
Registers a patch baseline for a patch group.

RegisterTargetWithMaintenanceWindow
Registers a target with a maintenance window.

RegisterTaskWithMaintenanceWindow
Adds a new task to a maintenance window.

RemoveTagsFromResource
Removes tag keys from the specified resource.

ResetServiceSetting

ResumeSession
Reconnects a session to an instance after it has been disconnected. Connections can be

SendAutomationSignal
Sends a signal to an Automation execution to change the current behavior or status of the

SendCommand
Runs commands on one or more managed instances.

StartAssociationsOnce
Use this API action to run an association immediately and only one time. This action can be

StartAutomationExecution
Initiates execution of an Automation document.

StartChangeRequestExecution
Creates a change request for Change Manager. The runbooks (Automation documents) specified in the

StartSession
Initiates a connection to a target (for example, an instance) for a Session Manager session. Returns a

StopAutomationExecution
Stop an Automation that is currently running.

TerminateSession
Permanently ends a session and closes the data connection between the Session Manager client and

UnlabelParameterVersion
Remove a label or labels from a parameter.

UpdateAssociation
Updates an association. You can update the association name and version, the document

UpdateAssociationStatus
Updates the status of the Systems Manager document associated with the specified instance.

UpdateDocument
Updates one or more values for an SSM document.

UpdateDocumentDefaultVersion
Set the default version of a document.

UpdateDocumentMetadata
Updates information related to approval reviews for a specific version of a document.

UpdateMaintenanceWindow
Updates an existing maintenance window. Only specified parameters are modified.

UpdateMaintenanceWindowTarget
Modifies the target of an existing maintenance window. You

UpdateMaintenanceWindowTask
Modifies a task assigned to a maintenance window. You can't change the task type, but you

UpdateManagedInstanceRole
Changes the Amazon Identity and Access Management (IAM) role that is assigned to the

UpdateOpsItem
Edit or change an OpsItem. You must have permission in AWS Identity and Access Management

UpdateOpsMetadata
Systems Manager calls this API action when you edit OpsMetadata in Application Manager.

UpdatePatchBaseline
Modifies an existing patch baseline. Fields not specified in the request are left

UpdateResourceDataSync
Update a resource data sync. After you create a resource data sync for a Region, you can't

UpdateServiceSetting